% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAGEr.R
\docType{class}
\name{CAGEr-class}
\alias{CAGEr-class}
\title{CAGEr objects}
\description{
The \emph{CAGEr} package provides one class of objects to load, contain and
process CAGE data: the \code{\link{CAGEexp}} class, introduced 2017, which is based on the
\code{\link[MultiAssayExperiment:MultiAssayExperiment]{MultiAssayExperiment::MultiAssayExperiment}} class.  In comparison with the original \code{CAGEset}
class (removed in 2021) \code{CAGEexp} objects benefit from a a more efficient data storage, using
\code{DataFrame}s of run-length-encoded (\code{Rle}) integers, allowing for the
loading and use of much larger transcriptome datasets.
}
\references{
Haberle V, Forrest ARR, Hayashizaki Y, Carninci P and Lenhard B
(\strong{2015}). \dQuote{CAGEr: precise TSS data retrieval and high-resolution
promoterome mining for integrative analyses.} \emph{Nucleic Acids Research},
43, pp. e51., \url{http://nar.oxfordjournals.org/content/43/8/e51}
}
