% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multicore.R
\name{CAGEr_Multicore}
\alias{CAGEr_Multicore}
\title{Multicore support in CAGEr}
\usage{
CAGEr_Multicore(useMulticore = FALSE, nrCores = NULL)
}
\description{
CAGEr is in the transition towards using the BiocParallel for
multicore parallelisation.  On Windows platforms, the multicore support
is disabled transparently, that is, attempts to use multiple cores are
silently ignored.
}
\author{
Charles Plessy
}
