% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportMethods.R
\name{coerceInBSgenome}
\alias{coerceInBSgenome}
\title{coerceInBSgenome}
\usage{
coerceInBSgenome(gr, genome)
}
\arguments{
\item{gr}{The genomic ranges to coerce.}

\item{genome}{The name of a BSgenome package, which must me installed,
or \code{NULL} to skip coercion.}
}
\value{
A GRanges object in which every range is guaranteed to be compatible
with the given BSgenome object.  The sequnames of the GRanges are also set
accordingly to the BSgenome.
}
\description{
A private (non-exported) function to discard any range that is
not compatible with the CAGEr object's BSgenome.
}
