% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{CCgranges2dataframe}
\alias{CCgranges2dataframe}
\alias{CCdataframe2granges}
\title{consensus Cluster convertors}
\usage{
CCgranges2dataframe(gr)

CCdataframe2granges(df)
}
\arguments{
\item{gr}{Consensus clusters in \code{GRanges} format.}

\item{df}{Consensus clusters in \code{data.frame} format.}
}
\description{
consensus Cluster convertors
}
\concept{df2granges converters}
