% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTSS.R
\docType{class}
\name{ConsensusClusters-class}
\alias{ConsensusClusters-class}
\alias{.ConsensusClusters}
\alias{ConsensusClusters}
\title{ConsensusClusters}
\description{
The \code{ConsensusClusters} class represents consensus clusters.
It is used internally by CAGEr for type safety.
}
\details{
Consensus clusters must not overlap, so that a single TSS in the
genome can only be attributed to a single cluster.
}
