% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\docType{methods}
\name{consensusClusters<-}
\alias{consensusClusters<-}
\alias{consensusClusters<-,CAGEset-method}
\alias{consensusClusters<-,CAGEexp-method}
\alias{consensusClustersSE<-}
\alias{consensusClustersSE<-,CAGEset,ANY-method}
\alias{consensusClustersSE<-,CAGEexp,RangedSummarizedExperiment-method}
\alias{consensusClustersGR<-}
\alias{consensusClustersGR<-,CAGEset-method}
\alias{consensusClustersGR<-,CAGEexp-method}
\title{Set consensus clusters from CAGEr objects}
\usage{
consensusClusters(object) <- value

\S4method{consensusClusters}{CAGEset}(object) <- value

\S4method{consensusClusters}{CAGEexp}(object) <- value

consensusClustersSE(object) <- value

\S4method{consensusClustersSE}{CAGEset,ANY}(object) <- value


  \S4method{consensusClustersSE}{CAGEexp,RangedSummarizedExperiment}(object) <- value

consensusClustersGR(object) <- value

\S4method{consensusClustersGR}{CAGEset}(object) <- value

\S4method{consensusClustersGR}{CAGEexp}(object) <- value
}
\arguments{
\item{object}{A \code{\link{CAGEr}} object.}

\item{value}{A \code{data.frame} of consensus clusters}
}
\description{
Set the information on consensus clusters in a \code{\link{CAGEr}}
object.
}
\details{
These setter methods are mostly for internal use, but are exported
in case they may be useful to advanced users.
}
\author{
Vanja Haberle

Charles Plessy
}
