% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\name{consensusClustersTpm}
\alias{consensusClustersTpm}
\alias{consensusClustersTpmDf}
\alias{consensusClustersTpmDF}
\title{Extracting consensus clusters tpm matrix from CAGEset object}
\usage{
consensusClustersTpm(object)
consensusClustersTpmDf(object)
consensusClustersTpmDF(object)
}
\arguments{
\item{object}{A CAGEr object.}
}
\value{
Returns normalized expression valuse of CAGE clusters across all samples.
\itemize{
  \item For \code{consensusClustersTpm}, a \code{matrix}.
  \item For \code{consensusClustersTpmDf}, a \code{data.frame}.
  \item For \code{consensusClustersTpmDF}, a \code{DataFrame}.
}
}
\description{
Extracts a table with normalized CAGE tag values for consensus
clusters across all samples from a \code{\link{CAGEr}} object.
}
\examples{
load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
clusters.tpm <- consensusClustersTpm(exampleCAGEset)
head(clusters.tpm)
consensusClustersTpmDF(exampleCAGEset)

}
\seealso{
\code{\link{consensusClusters}}
}
\author{
Vanja Haberle
}
