% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_r_docs.R
\name{convert_r_docs}
\alias{convert_r_docs}
\title{Convert R Docs to Data}
\usage{
convert_r_docs(pkg)
}
\arguments{
\item{pkg}{a string indicating the package to convert}
}
\value{
a dataframe of prompts and responses
}
\description{
This function converts R documentation into prompts and responses
}
