% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationMethods.R
\name{correlationMatrix}
\alias{correlationMatrix}
\alias{correlationMatrix,CAGEexp-method}
\alias{correlationMatrix,SummarizedExperiment-method}
\alias{correlationMatrix,DataFrame-method}
\alias{correlationMatrix,data.frame-method}
\alias{correlationMatrix,matrix-method}
\title{Pairwise correlations of CAGE signal}
\usage{
correlationMatrix(
  object,
  what = c("CTSS", "consensusClusters"),
  values = c("raw", "normalized"),
  samples = "all",
  method = "pearson",
  tagCountThreshold = 1,
  applyThresholdBoth = FALSE
)

\S4method{correlationMatrix}{CAGEexp}(
  object,
  what = c("CTSS", "consensusClusters"),
  values = c("raw", "normalized"),
  samples = "all",
  method = "pearson",
  tagCountThreshold = 1,
  applyThresholdBoth = FALSE
)

\S4method{correlationMatrix}{SummarizedExperiment}(
  object,
  what = c("CTSS", "consensusClusters"),
  values = c("raw", "normalized"),
  samples = "all",
  method = "pearson",
  tagCountThreshold = 1,
  applyThresholdBoth = FALSE
)

\S4method{correlationMatrix}{DataFrame}(
  object,
  what = c("CTSS", "consensusClusters"),
  values = c("raw", "normalized"),
  samples = "all",
  method = "pearson",
  tagCountThreshold = 1,
  applyThresholdBoth = FALSE
)

\S4method{correlationMatrix}{data.frame}(
  object,
  what = c("CTSS", "consensusClusters"),
  values = c("raw", "normalized"),
  samples = "all",
  method = "pearson",
  tagCountThreshold = 1,
  applyThresholdBoth = FALSE
)

\S4method{correlationMatrix}{matrix}(
  object,
  what = c("CTSS", "consensusClusters"),
  values = c("raw", "normalized"),
  samples = "all",
  method = "pearson",
  tagCountThreshold = 1,
  applyThresholdBoth = FALSE
)
}
\arguments{
\item{object}{A \code{\link{CAGEr}} object or (only for
\code{plotCorrelation2}) a \code{\link{SummarizedExperiment}} or an
expression table as a \code{\link{DataFrame}}, \code{\link{data.frame}} or
\code{\link{matrix}} object.}

\item{what}{The clustering level to be used for calculating
correlations. Can be either \code{"CTSS"} to use individual TSSs or
\code{"consensusClusters"} to use consensus clusters, \emph{i.e.} entire
promoters.  Ignored for anything else than \code{CAGEr} objects.}

\item{values}{Use either \code{"raw"} (default) or \code{"normalized"} CAGE
signal. Ignored for plain expression tables.}

\item{samples}{Character vector indicating which samples to use. Can be
either \code{"all"} to select all samples in a \code{CAGEr} object, or a
subset of valid sample labels as returned by the
\code{\link{sampleLabels}} function.}

\item{method}{A character string indicating which correlation coefficient
should be computed.  Passed to \code{cor} function.  Can be one of
\code{"pearson"}, \code{"spearman"}, or \code{"kendall"}.}

\item{tagCountThreshold}{Only TSSs with tag count \code{>= tagCountThreshold}
in either one (\code{applyThresholdBoth = FALSE}) or both samples
(\code{applyThresholdBoth = TRUE}) are used to calculate
correlation.}

\item{applyThresholdBoth}{See \code{tagCountThreshold} above.}
}
\value{
Returns a \code{matrix} of pairwise correlations between
selected samples.
}
\description{
Calculates the pairwise correlation between samples
}
\details{
The same matrix is returned from \code{plotCorrelation2}

\code{SummarizedExperiment} objects are expected to contain raw tag counts
in a \dQuote{counts} assay and the normalized expression scores in a
\dQuote{normalized} assay.

Avoid using large \code{matrix} objects as they are coerced to
\code{DataFrame} class without special care for efficiency.
}
\examples{

correlationMatrix(exampleCAGEexp, what = "consensusClusters", value = "normalized")

}
\author{
Vanja Haberle

Charles Plessy

Katalin Ferenc
}
\concept{CAGEr functions}
