% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_content.R
\name{create_content}
\alias{create_content}
\title{Create Content for Tuning}
\usage{
create_content(func, args_to_use)
}
\arguments{
\item{func}{a string indicating the function name that could be called in R}

\item{args_to_use}{arguments to use in the function}
}
\value{
a JSON string with a prompt and the function to be written
}
\description{
This function converts a function into content usable for
an LLM tuning.
}
