% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringMethods.R
\docType{methods}
\name{.ctssAggregateAndSum}
\alias{.ctssAggregateAndSum}
\alias{.ctssAggregateAndSum,data.table-method}
\title{Aggregate identical CTSS and sum their scores.}
\usage{
.ctssAggregateAndSum(ctss)

\S4method{.ctssAggregateAndSum}{data.table}(ctss)
}
\description{
Private function using  \code{data.table} objects to preform grouping
operations at a high performance.  These functions use \emph{non-standard evaluation}
in a context that raises warnings in \code{R CMD check}.  By separating these functions
from the rest of the code, I hope to make the workarounds easier to manage.
}
\examples{
ctssDT <- data.table( chr       = c("chr1", "chr1", "chr1", "chr2")
                    , pos       = c(1     , 1     , 2     , 1     )
                    , strand    = c("+"   , "+"   , "-"   , "-"   )
                    , tag_count = c(1     , 1     , 1     , 1     ))
ctssDT
.ctssAggregateAndSum(ctssDT)
}
