% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\docType{methods}
\name{CTSSclusteringMethod}
\alias{CTSSclusteringMethod}
\alias{CTSSclusteringMethod,CAGEset}
\alias{CTSSclusteringMethod,CAGEexp}
\alias{CTSSclusteringMethod,CAGEset-method}
\alias{CTSSclusteringMethod,CAGEexp-method}
\title{Extracting CTSS clustering method from CAGEr objects.}
\usage{
CTSSclusteringMethod(object)

\S4method{CTSSclusteringMethod}{CAGEset}(object)

\S4method{CTSSclusteringMethod}{CAGEexp}(object)
}
\arguments{
\item{object}{A CAGEset or CAGEexp object.}
}
\value{
Returns a label of the method used for CTSS clustering.
}
\description{
Extracts the label of the method used for CTSS clustering into tag
clusters from \code{\link{CAGEset}} and \code{\link{CAGEexp}} objects.
}
\examples{
load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
CTSSclusteringMethod(exampleCAGEset)

}
\seealso{
\code{\link{clusterCTSS}}

Other CAGEr accessor methods: \code{\link{CTSScoordinates}},
  \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CTSSnormalizedTpm}},
  \code{\link{CTSStagCountTable}},
  \code{\link{CTSStagCount}}, \code{\link{GeneExpDESeq2}},
  \code{\link{GeneExpSE}}, \code{\link{genomeName}},
  \code{\link{inputFilesType}}, \code{\link{inputFiles}},
  \code{\link{librarySizes}}, \code{\link{sampleLabels}},
  \code{\link{tagClusters}}
}
\author{
Vanja Haberle
}
