% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAGEfightR.R
\name{.export_tag_counts}
\alias{.export_tag_counts}
\title{Helper function to export tag count data into summarized experiment object with StitchedGPos.}
\usage{
.export_tag_counts(object = "CAGEexp")
}
\arguments{
\item{object}{A \code{CAGEexp} object}
}
\value{
A \code{SummarizedExperiment} object.
}
\description{
Helper function to export tag count data into summarized experiment object with StitchedGPos.
}
\seealso{
Other CAGEr accessor methods: 
\code{\link{CTSScoordinatesGR}()},
\code{\link{CTSScumulativesTagClusters}()},
\code{\link{CTSSnormalizedTpmDF}()},
\code{\link{CTSStagCountDF}()},
\code{\link{GeneExpDESeq2}()},
\code{\link{GeneExpSE}()},
\code{\link{consensusClustersGR}()},
\code{\link{exportNormalizedCTSS}()},
\code{\link{expressionClasses}()},
\code{\link{filteredCTSSidx}()},
\code{\link{genomeName}()},
\code{\link{inputFiles}()},
\code{\link{inputFilesType}()},
\code{\link{librarySizes}()},
\code{\link{sampleLabels}()},
\code{\link{seqNameTotalsSE}()},
\code{\link{tagClustersGR}()}

Other CAGEr normalized data functions: 
\code{\link{exportNormalizedCTSS}()}
}
\concept{CAGEr accessor methods}
\concept{CAGEr normalized data functions}
