% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAGEexp.R
\docType{data}
\name{exampleCAGEexp}
\alias{exampleCAGEexp}
\title{Example CAGEexp object.}
\format{A \code{\link{CAGEexp}} object.}
\description{
Lazy-loaded example CAGEexp object, containing most of the CAGEr data
structures created with the CAGEr modifier functions.
}
\examples{
# The example commands in the CAGEexp help page were used to create the example
# CAGEexp object used below.
\dontrun{
exampleCAGEexp <- exampleCAGEexp[,c(5, 2, 1, 3, 4)] # Non-aplhabetic order may help catch bugs
CTSStagCountSE(exampleCAGEexp) <- CTSStagCountSE(exampleCAGEexp)[1:5000,]  # Slim the object
summariseChrExpr(exampleCAGEexp)
annotateCTSS(exampleCAGEexp, exampleZv9_annot)
CTSStoGenes(exampleCAGEexp)
normalizeTagCount(exampleCAGEexp)
clusterCTSS(exampleCAGEexp)
cumulativeCTSSdistribution(exampleCAGEexp, "tagClusters")
quantilePositions(exampleCAGEexp, "tagClusters")
aggregateTagClusters(exampleCAGEexp)
annotateConsensusClusters(exampleCAGEexp, exampleZv9_annot)
cumulativeCTSSdistribution(exampleCAGEexp, "consensusClusters")
quantilePositions(exampleCAGEexp, "consensusClusters")
save(exampleCAGEexp, file = "data/exampleCAGEexp.RData", compress = "xz")
}
}
