\name{exampleCAGEset}
\docType{data}
\alias{exampleCAGEset}
\title{Example CAGEset object with zebrafish CAGE data}
\description{
This is a \code{\link{CAGEset}} object that contains CAGE data for part of chromosome 13 from five different zebrafish (\emph{Danio rerio}) samples. It is intended to be used as an input data in running examples from \code{\link{CAGEr}} package help pages.
}
\usage{data(exampleCAGEset)}
\format{A \code{\link{CAGEset}} object}
\keyword{datasets}