% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAGEfightR.R
\name{exportNormalizedCTSS}
\alias{exportNormalizedCTSS}
\alias{exportNormalizedCTSS,CAGEexp-method}
\title{Export of normalized CTSS in an object that can be used as input to CAGEfightR.}
\usage{
exportNormalizedCTSS(object)

\S4method{exportNormalizedCTSS}{CAGEexp}(object)
}
\arguments{
\item{object}{A \code{CAGEexp} object with CTSS values}
}
\value{
A \code{SummarizedExperiment} object with StitchedGPos values.
}
\description{
An export function for integration of normalized values into CAGEfightR.
}
\note{
At the moment the conversion is expensive as it goes from \code{DataFrame}
of \code{Rle} to \code{data.frame} to \code{matrix}.
}
\examples{
exportNormalizedCTSS(exampleCAGEexp)
}
\seealso{
Other CAGEr accessor methods: 
\code{\link{.export_tag_counts}()},
\code{\link{CTSScoordinatesGR}()},
\code{\link{CTSScumulativesTagClusters}()},
\code{\link{CTSSnormalizedTpmDF}()},
\code{\link{CTSStagCountDF}()},
\code{\link{GeneExpDESeq2}()},
\code{\link{GeneExpSE}()},
\code{\link{consensusClustersGR}()},
\code{\link{expressionClasses}()},
\code{\link{filteredCTSSidx}()},
\code{\link{genomeName}()},
\code{\link{inputFiles}()},
\code{\link{inputFilesType}()},
\code{\link{librarySizes}()},
\code{\link{sampleLabels}()},
\code{\link{seqNameTotalsSE}()},
\code{\link{tagClustersGR}()}

Other CAGEr normalized data functions: 
\code{\link{.export_tag_counts}()}
}
\concept{CAGEr accessor methods}
\concept{CAGEr normalized data functions}
