% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportMethods.R
\docType{methods}
\name{exportToBed}
\alias{exportToBed}
\alias{exportToBed,CAGEr-method}
\title{Create BED tracks of TSSs and clusters of TSSs}
\usage{
exportToBed(object, what = c("CTSS", "tagClusters", "consensusClusters"),
  qLow = NULL, qUp = NULL, colorByExpressionProfile = FALSE,
  oneFile = TRUE)

\S4method{exportToBed}{CAGEr}(object, what = c("CTSS", "tagClusters",
  "consensusClusters"), qLow = NULL, qUp = NULL,
  colorByExpressionProfile = FALSE, oneFile = TRUE)
}
\arguments{
\item{object}{A \code{\link{CAGEr}} object.}

\item{what}{Which elements should be exported to BED track.  \code{CTSS} to export
individual CTSSs,  \code{tagClusters} to export tag clusters or \code{consensusClusters}
to export consensus clusters.}

\item{qLow, qUp}{Position of which "lower" (resp. "upper") quantile should be
used as 5' (resp. 3') boundary of the filled block in gene-like
representation of the cluster.  Default value \code{NULL} uses start (resp. end)
position of the cluster.  Ignored when \code{what = "CTSS"}.}

\item{colorByExpressionProfile}{Logical, should blocks be colored in the
color of their corresponding expression class.  Ignored when
\code{what = "tagClusters"}.}

\item{oneFile}{Logical, should all CAGE datasets be exported as individual
tracks into the same BED file (\code{TRUE}) or into separate BED files (\code{FALSE}).
Ignored when \code{what = "CTSS"}, which by default produces only one track.}
}
\value{
Creates BED file(s) in the working directory.
}
\description{
Creates BED file(s) with track(s) of individual CTSSs, tag
clusters or consensus clusters.  CTSSs and consensus clusters can be
optionally colored in the color of their expression class.  \emph{Tag clusters}
and \emph{consensus clusters} can be displayed in a gene-like representation with
a line showing full span on the cluster, filled block showing interquantile
range and a thick box denoting position of the dominant (most frequently
used TSS.
}
\details{
The BED representations of \emph{CTSSs}, \emph{tag cluster} and
\emph{consensus clusters} can be directly visualised in the ZENBU or UCSC Genome
Browsers.

When \code{what = "CTSS"}, one BED file with single track of 1 bp blocks
representing all detected CTSSs (in all CAGE samples) is created.  CTSSs can
be colored according to their expression class (provided the expression
profiling of CTSSs was done by calling \code{\link{getExpressionProfiles}} function).
Colors of expression classes match the colors in which they are shown in the
plot returned by the \code{\link{plotExpressionProfiles}} function.  For
\code{colorByExpressionProfile = FALSE}, CTSSs included in the clusters are
shown in black and CTSSs that were filtered out in gray.

When \code{what = "tagClusters"}, one track per CAGE dataset is created, which can
be exported to a single BED file (by setting \code{oneFile = TRUE}) or separate
BED files (\code{FALSE}).  If no quantile boundaries were provided (\code{qLow} and
\code{qUp} are \code{NULL}, TCs are represented as simple blocks showing the full
span of TC fromthe start to the end.  Setting \code{qLow} and/or \code{qUp} parameters
to a value of the desired quantile creates a gene-like representation with a
line showing full span of the TC, filled block showing specified
interquantile range and a thick 1 bp block denoting position of the dominant
(most frequently used) TSS.  All TCs in one track (one CAGE dataset) are
shown in the same color.

When \code{what = "consensusClusters"} \emph{consensus clusters} are exported to BED
file.  Since there is only one set of consensus clusters common to all CAGE
datasets, only one track is created in case of a simple representation.  This
means that when \code{qLow = NULL} and \code{qUp = NULL} one track with blocks showing
the full span of consensus cluster from the start to the end is created.
However, the distribution of the CAGE signal within consensus cluster can be
different in different CAGE samples, resulting in different positions of
quantiles and dominant TSS.  Thus, when \code{qLow} and/or \code{qUp} parameters
are set to a value of the desired quantile, a separate track with a gene-like
representation is created for every CAGE dataset.  These tracks can be
exported to a single BED file (by setting \code{oneFile = TRUE}) or separate
BED files (by setting \code{oneFile = FALSE}).  The gene-like representation is
analogous to the one described above for the TCs.  In all cases consensus
clusters can be colored according to their expression class (provided the
expression profiling of consensus clusters was done by calling
\code{getExpressionProfiles} function).  Colors of expression classes match the
colors in which they are shown in the plot returned by the
\code{plotExpressionProfiles} function.  For \code{colorByExpressionProfile = FALSE}
all consensus clusters are shown in black.
}
\examples{

### exporting CTSSs colored by expression class
exportToBed(object = exampleCAGEset, what = "CTSS", colorByExpressionProfile = TRUE)

### exporting tag clusters in gene-like representation
exportToBed( object = exampleCAGEset, what = "tagClusters"
           , qLow = 0.1, qUp = 0.9, oneFile = TRUE)
           
exportToBed( object = exampleCAGEexp, what = "tagClusters"
           , qLow = 0.1, qUp = 0.9, oneFile = TRUE)

}
\seealso{
Other CAGEr export functions: \code{\link{exportCTSStoBedGraph}}
}
\author{
Vanja Haberle
}
\concept{CAGEr export functions}
