% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportMethods.R
\docType{data}
\name{FANTOM5humanSamples}
\alias{FANTOM5humanSamples}
\title{FANTOM5 human samples}
\format{
A \code{data.frame} with \code{sample}, \code{type}, \code{description}, \code{library_id}
and \code{data_url} columns.
}
\usage{
FANTOM5humanSamples
}
\description{
Lazy-loaded \code{data.frame} object, containing information about FANTOM5
libraries.  Its use is described in more details in the vignette
\dQuote{Use of CAGE resources with CAGEr}.
}
\seealso{
Other FANTOM data: 
\code{\link{FANTOM5mouseSamples}}
}
\concept{FANTOM data}
\keyword{datasets}
