% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizationFunctions.R
\name{fitPowerLaw}
\alias{fitPowerLaw}
\alias{fitPowerLaw,CAGEexp-method}
\alias{fitPowerLaw,DataFrame-method}
\alias{fitPowerLaw,GRangesList-method}
\alias{fitPowerLaw,GRanges-method}
\alias{fitPowerLaw,Rle-method}
\alias{fitPowerLaw,numeric-method}
\title{Fit a power law to the CAGE data}
\usage{
fitPowerLaw(x, fitInRange = c(10, 1000))

\S4method{fitPowerLaw}{CAGEexp}(x, fitInRange = c(10, 1000))

\S4method{fitPowerLaw}{DataFrame}(x, fitInRange = c(10, 1000))

\S4method{fitPowerLaw}{GRangesList}(x, fitInRange = c(10, 1000))

\S4method{fitPowerLaw}{GRanges}(x, fitInRange = c(10, 1000))

\S4method{fitPowerLaw}{Rle}(x, fitInRange = c(10, 1000))

\S4method{fitPowerLaw}{numeric}(x, fitInRange = c(10, 1000))
}
\arguments{
\item{x}{The object containing the data.}

\item{fitInRange}{The range in which to fit the power law.}
}
\value{
Returns the slope (\code{plSlope}) and the intercept (\code{plInt}) of the
fitted power-law distribution.  If the input is a single sample, function
returns a numeric vector.  If the input is multiple samples, the function
returns a \code{DataFrame}.  If the input is a whole \emph{CAGEexp} object, the
function returns the object with information added to its \code{colData}.
}
\description{
Function that fits power-law distribution to reverse cumulative of given
values.  Fitting is done using only a specified range.
}
\examples{
exampleCAGEexp |> fitPowerLaw() |> colData()
exampleCAGEexp |> CTSStagCountDF() |> fitPowerLaw()
exampleCAGEexp |> CTSStagCountGR('all') |> fitPowerLaw()
exampleCAGEexp |> CTSStagCountGR(1) |> fitPowerLaw()
exampleCAGEexp |> CTSStagCountGR(1) |> score() |> fitPowerLaw()
exampleCAGEexp |> CTSStagCountGR(1) |> score() |> decode() |> fitPowerLaw()
}
\author{
Vanja Haberle

Charles Plessy
}
