% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportMethods.R
\name{import.bigwig}
\alias{import.bigwig}
\title{Read in BigWig files to CAGEexp object}
\usage{
import.bigwig(genome, filepath)
}
\arguments{
\item{genome}{the name of the reference genome (bsgenome)}

\item{filepath}{list of input bigwig files with full path
The filenames should include a "_str1" and "_str2" substring after the sample name
The sampe name will be the prefix before these substrings
the filepath in the CAGEexp object will only include the bw corresponding to str1}
}
\value{
a CAGEexp object
}
\description{
Read in BigWig files to CAGEexp object
}
\examples{
import.bigwig(
 genome="BSgenome.Drerio.UCSC.danRer7",
 filepath = system.file("extdata", "NCig10061_subsampled_str1.Signal.Unique.str1.out.wig.bw", package = "CAGEr")
 )
}
\seealso{
Other loadFileIntoGPos: 
\code{\link{bam2CTSS}()},
\code{\link{import.CTSS}()},
\code{\link{import.bam}()},
\code{\link{import.bam.ctss}()},
\code{\link{import.bedCTSS}()},
\code{\link{import.bedScore}()},
\code{\link{import.bedmolecule}()},
\code{\link{loadFileIntoGPos}()},
\code{\link{moleculesGR2CTSS}()}
}
\author{
Katalin Ferenc

Damir Baranasic
}
\concept{loadFileIntoGPos}
