% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportMethods.R
\name{import.CAGEscanMolecule}
\alias{import.CAGEscanMolecule}
\title{import.CAGEscanMolecule}
\usage{
import.CAGEscanMolecule(filepath)
}
\arguments{
\item{filepath}{The path to the \dQuote{molecule} file.}
}
\description{
Imports a CAGEscan \dQuote{molecule} file in a \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}} object
}
\examples{
# TODO import.CAGEscanMolecule(system.file("extdata", "example.molecule.txt", package = "CAGEr"))

}
\seealso{
parseCAGEscanBlocksToGrangeTSS
}
