% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergingMethods.R
\docType{methods}
\name{mergeSamples}
\alias{mergeSamples}
\alias{mergeSamples,CAGEset,numeric-method}
\alias{mergeSamples,CAGEexp,ANY-method}
\title{Merging CAGE datasets (samples)}
\usage{
mergeSamples(object, mergeIndex, mergedSampleLabels)

\S4method{mergeSamples}{CAGEset,numeric}(object, mergeIndex, mergedSampleLabels)

\S4method{mergeSamples}{CAGEexp,ANY}(object, mergeIndex, mergedSampleLabels)
}
\arguments{
\item{object}{A \code{\link{CAGEr}} object.}

\item{mergeIndex}{Integer vector specifying which experiments should be merged.  Must be the
same length as the number of datasets in the CAGEr object.  See Details.}

\item{mergedSampleLabels}{Labels for the merged datasets.  Must be the same lenght as the number
of unique values in \code{mergeIndex}.  See Details.}
}
\value{
The slots \code{sampleLabels}, \code{librarySizes} and \code{tagCountMatrix} of the
provided \code{\link{CAGEr}} object will be updated with the information on merged CAGE datasets
and will replace the previous information on individual CAGE datasets.  All further slots with
downstream information will be reset.
}
\description{
Merges individual CAGE datasets (experiments, samples) within the CAGEr object
into specified groups.
}
\details{
This function merges CAGE datasets within the CAGEr object at the level of CTSS tag
counts, \emph{i.e.} tag counts of individual CTSS are summed over a group of datasets that are
being merged.  After merging, all other slots in the CAGEr object will be reset and any previous
data for individual experiments will be removed.

\code{mergeIndex} controls which datasets will be merged.  It is an integer vector that assigns
a value to each dataset in the CAGEr object in the same order as they are returned by
\code{sampleLabels(object)}.  Datasets with the same value in the \code{mergeIndex} will be
merged.  For example, if there are 8 CAGE datasets in the CAGEr object and \code{mergeIndex =
c(1,1,2,2,3,2,4,4)}, this will merge a) samples 1 and 2 b) samples 3, 4 and 6 c) samples 7 and
8, and it will leave sample 5 as it is, resulting in 4 final merged datasets.

Labels provided in \code{mergedSampleLabels} will be assigned to merged datasets in the ascending
order of \code{mergeIndex} values, \emph{i.e.} first label will be assigned to a dataset created
by merging datasets labeled with lowest \code{mergeIndex} value (in this case \code{1}),
\emph{etc}.
}
\examples{
mergeSamples( exampleCAGEset
            , mergeIndex = c(1,1,2)
            , mergedSampleLabels = c("mergedSample1", "mergedSample2"))
exampleCAGEset

mergeSamples( exampleCAGEexp
            , mergeIndex = c(3,2,4,4,1)
            , mergedSampleLabels = c("zf_unfertilized", "zf_high", "zf_30p_dome", "zf_prim6"))
ce

}
\author{
Vanja Haberle
}
