% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Richness.R
\name{plot.hanabi}
\alias{plot.hanabi}
\title{plot.hanabi}
\usage{
plot.hanabi(x, alpha = 0.5, col = "black", xlab = "Total counts",
  ylab = "Unique features", main = "Hanabi plot", pch = 1, ...)
}
\arguments{
\item{x}{The hanabi object to plot.}

\item{alpha}{The alpha transparency of the plot lines.}

\item{col}{A vector indicating a color per sample (or a vector that
can be recycled that way).}

\item{xlab}{Horizontal axis label.}

\item{ylab}{Vertical axis label.}

\item{main}{Plot title.}

\item{pch}{Plot character at the tip of the lines.}

\item{...}{other arguments passed to the generic plot function.}
}
\description{
Plotting Hanabi objects
}
\seealso{
Other CAGEr richness functions: \code{\link{hanabiPlot}},
  \code{\link{hanabi}}, \code{\link{points.hanabi}}
}
\author{
Charles Plessy
}
