% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportMethods.R
\docType{methods}
\name{plotInterquantileWidth}
\alias{plotInterquantileWidth}
\alias{plotInterquantileWidth,CAGEr-method}
\title{Plot distribution of interquantile width}
\usage{
plotInterquantileWidth(object, clusters, tpmThreshold = 5, qLow = 0.1,
  qUp = 0.9, xlim = c(0, 150), ...)

\S4method{plotInterquantileWidth}{CAGEr}(object, clusters, tpmThreshold = 5,
  qLow = 0.1, qUp = 0.9, xlim = c(0, 150), ...)
}
\arguments{
\item{object}{A \code{\link{CAGEset}} object}

\item{clusters}{Which clusters should be used.  Can be either \code{"tagClusters"} to plot
distribution of interquantile width of tag clusters or \code{"consensusClusters"}
to plot distribution of interquantile width of consensus clusters.}

\item{tpmThreshold}{Only clusters with normalized signal \code{>= tpmThreshold} will be
included in the histogram.}

\item{qLow, qUp}{Position of which "lower" (resp. "upper") quantile should be used as 5'
(resp. 3') boundary.  See Details.}

\item{xlim}{The x axis limits of the plot.}

\item{...}{Additional arguments passed to \code{plot()} function, such as \code{ylim},
\emph{etc.}.}
}
\description{
Plots histograms showing distribution of interquantile width of tag clusters or
consensus clusters in each CAGE dataset.
}
\details{
Interquantile width is a width (in base-pairs) of the central part of the genomic
region (bounded by the positions of specified \code{qLow} and \code{qUp} quantiles) that
contains \code{>= (qUp - qLow) * 100\%} of the CAGE signal.  Positions of specified quantiles
within each cluster have to be calculated beforehand by calling \code{\link{quantilePositions}}
function.  Interquantile width is a more robust measure of the promoter width than the total
span of the region, because it takes into account the magnitude of the expression in the
region.
}
\examples{
load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
plotInterquantileWidth( object = exampleCAGEset, clusters = "tagClusters"
                      , tpmThreshold = 50, qLow = 0.1, qUp = 0.9)

}
\seealso{
Other CAGEr export functions: \code{\link{exportCTSStoBedGraph}},
  \code{\link{exportToBed}}

Other CAGEr clusters functions: \code{\link{CTSScumulativesTagClusters}},
  \code{\link{aggregateTagClusters}},
  \code{\link{clusterCTSS}},
  \code{\link{consensusClustersDESeq2}},
  \code{\link{consensusClusters}},
  \code{\link{cumulativeCTSSdistribution}},
  \code{\link{quantilePositions}},
  \code{\link{tagClusters}}
}
\author{
Vanja Haberle
}
