% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Richness.R
\name{points.hanabi}
\alias{points.hanabi}
\title{points.hanabi}
\usage{
\method{points}{hanabi}(x, ...)
}
\arguments{
\item{x}{The hanabi object.}

\item{...}{Other parameters passed to the generic points function}
}
\description{
Add a final point in hanabi plots.
}
\details{
Will only add a point for the final, non-subsampled value of each
sample of in a hanabi object.
}
\seealso{
Other CAGEr richness functions: \code{\link{hanabiPlot}},
  \code{\link{hanabi}}, \code{\link{plot.hanabi}}
}
\author{
Charles Plessy
}
