% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAGEfightR.R
\name{quickEnhancers,CAGEexp-method}
\alias{quickEnhancers,CAGEexp-method}
\alias{quickEnhancers}
\title{Identify and quantify enhancers.}
\usage{
\S4method{quickEnhancers}{CAGEexp}(object)
}
\arguments{
\item{object}{A \code{CAGEexp} object}
}
\value{
A \code{RangedSummarizedExperiment} object.  See the example below on
how to attach it to the experiment list of a \code{CAGEexp} object.
}
\description{
A convenient wrapper to the function \code{\link[CAGEfightR:quickEnhancers]{CAGEfightR::quickEnhancers()}}.
}
\details{
The \code{CAGEr} object will be converted to a format similar to the output
of \code{\link[CAGEfightR:quantifyCTSSs]{CAGEfightR::quantifyCTSSs()}}, and then passed to the \code{quickEnhancers}
function.
}
\note{
At the moment the conversion is expensive as it goes from \code{DataFrame}
of \code{Rle} to \code{data.frame} to \code{matrix}.
}
\examples{
# The CAGEfightR package needs to be installed before running the command.
\dontrun{
enhancers <- quickEnhancers(ce)
ce <- c(enhancers = enhancers, ce)
}

}
\concept{CAGEfightR}
