% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotations.R
\name{ranges2genes}
\alias{ranges2genes}
\title{ranges2genes}
\usage{
ranges2genes(ranges, genes)
}
\arguments{
\item{ranges}{Genomics Ranges object, for example extracted from a
RangedSummarizedExperiment object with the \code{rowRanges}
command.}

\item{genes}{A \code{\link{GRanges}} object containing \code{gene_name} metadata.}
}
\value{
A character vector of same length as the GRanges object, indicating
        one gene symbol or a comma-separated list of gene symbols for each
        range.
}
\description{
Assign gene symbol(s) to Genomic Ranges.
}
\examples{
ce <- readRDS(system.file(package = "CAGEr", "extdata/CAGEexp.rds"))
gff <- readRDS(system.file("extdata/Zv9_annot.rds", package = "CAGEr"))
CTSScoordinatesGR(ce)$genes <- ranges2genes(CTSScoordinatesGR(ce), gff)

}
\seealso{
\code{\link{CTSScoordinatesGR}}, \code{\link{Zv9_annot}}

Other CAGEr annotation functions: \code{\link{annotateCTSS}},
  \code{\link{plotAnnot}}, \code{\link{ranges2annot}}
}
\author{
Charles Plessy
}
