% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResetMethods.R
\name{resetCAGEexp}
\alias{resetCAGEexp}
\alias{resetCAGEexp,CAGEexp-method}
\title{Reset a CAGEexp object}
\usage{
resetCAGEexp(object)

\S4method{resetCAGEexp}{CAGEexp}(object)
}
\arguments{
\item{object}{A \code{CAGEexp} object}
}
\value{
Returns a \code{CAGEexp} object, which contains a non-normalised
\code{tagCountMatrix} experiment.
}
\description{
Removes all data but the raw CTSS counts and coordinates from a \code{\link{CAGEexp}}
object.  Useful after removing samples.
}
\examples{
resetCAGEexp(exampleCAGEexp)

}
\seealso{
Other CAGEr object modifiers: 
\code{\link{CTSStoGenes}()},
\code{\link{CustomConsensusClusters}()},
\code{\link{aggregateTagClusters}()},
\code{\link{annotateCTSS}()},
\code{\link{clusterCTSS}()},
\code{\link{cumulativeCTSSdistribution}()},
\code{\link{getCTSS}()},
\code{\link{normalizeTagCount}()},
\code{\link{quantilePositions}()},
\code{\link{quickEnhancers}()},
\code{\link{summariseChrExpr}()}
}
\author{
Charles Plessy
}
\concept{CAGEr object modifiers}
