% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetMethods.R
\docType{methods}
\name{setColors}
\alias{setColors}
\alias{setColors,CAGEr-method}
\title{Setting colors for samples}
\usage{
setColors(object, colors = NULL)

\S4method{setColors}{CAGEr}(object, colors = NULL)
}
\arguments{
\item{object}{A \code{\link{CAGEr}} object.}

\item{colors}{A character vector of valid color names.  For a complete list of valid color
names see the \code{\link{colors}} function.  Alternatively, it can be a
character vector of colors specified in hexadecimal format (\emph{e.g.}
"#FF0000" for red).  Number of provided colors must match the number of
samples in the CAGEset object.  Provided colors are assigned to samples
according to their ordering in the CAGEr object, \emph{i.e} in the order
they are returned by \code{\link{sampleLabels}} function.}
}
\value{
Assigns one color to each sample in the CAGEr object by setting them as a name
attribute of the \code{sampleLabels} slot.
}
\description{
Assigns one color to each sample in the CAGEr object.  These colors are
used in various plots and exported tracks to consistently represent corresponding samples.
}
\examples{

load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
sampleLabels(exampleCAGEset)
setColors(exampleCAGEset, colors = c("darkred", "navy", "forestgreen"))
sampleLabels(exampleCAGEset)

ce <- readRDS(system.file(package = "CAGEr", "extdata/CAGEexp.rds"))
sampleLabels(ce)
setColors(ce, c("red", "darkgreen", "blue", "grey", "black"))
sampleLabels(ce)

}
\seealso{
\code{\link{sampleLabels}}

Other CAGEr setter methods: \code{\link{genomeName}},
  \code{\link{inputFilesType}}, \code{\link{inputFiles}},
  \code{\link{sampleLabels}}
}
\author{
Vanja Haberle
}
