\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,CAGEset-method}
\title{Methods for function \code{show}}
\description{
Methods for function \code{show}.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "CAGEset")}}{
Displays a CAGEset object in a user-friendly way, giving an overview of its content.
}
}}
\keyword{methods}
