% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringFunctions.R
\name{tagClusterConvertors}
\alias{tagClusterConvertors}
\alias{CCgranges2dataframe}
\alias{TCgranges2dataframe}
\alias{CCdataframe2granges}
\alias{TCdataframe2granges}
\title{Private functions to convert TC formats}
\usage{
TCgranges2dataframe(gr)

TCdataframe2granges(df)
}
\arguments{
\item{gr}{Consensus clusters in \code{GRanges} format.}

\item{df}{Consensus clusters in \code{data.frame} format.}
}
\details{
Interconvert tag clusters (TC) formats used in classes CAGEset
(\code{data.frame}) and CAGEexp (\code{GRanges}).
}
\examples{
load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
df <- tagClusters(object, 1)
head(df)
gr <- TCdataframe2granges(df)
gr
head(TCgranges2dataframe(gr))
# No round-trip because start and end were not integer in df.
if (! identical(df, TCgranges2dataframe(gr))) stop("No round-trip between TCdataframe2granges and TCgranges2dataframe")
if (! all(df == TCgranges2dataframe(gr))) stop("No round-trip between TCdataframe2granges and TCgranges2dataframe")
}
\seealso{
Other df2granges converters: \code{\link{consensusClusterConvertors}}
}
