% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTSS.R
\docType{class}
\name{TagClusters-class}
\alias{TagClusters-class}
\alias{.TagClusters}
\alias{TagClusters}
\title{TagClusters}
\description{
TagClusters
}
\details{
The \code{TagClusters} class represents tag clusters.
It is used internally by CAGEr for type safety.
}
