% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R, R/SetMethods.R
\name{tagClustersQuantile}
\alias{tagClustersQuantile}
\alias{tagClustersQuantileLow}
\alias{tagClustersQuantileUp}
\alias{`tagClustersQuantileLow<-`}
\alias{tagClustersQuantileUp}
\title{Quantile metadata stored in CAGEr objects.}
\usage{
tagClustersQuantileLow(object, samples = NULL)

tagClustersQuantileUp(object, samples = NULL)

tagClustersQuantileLow(object) <- value

tagClustersQuantileUp(object) <- value
}
\arguments{
\item{object}{A \code{\link{CAGEr}} object.}

\item{samples}{Sample name(s), number(s) or \code{NULL} (default) for all samples.}

\item{value}{A list (one entry per sample) of data frames with multiple columns:
\code{cluster} for the cluster ID, and then \code{q_0.n} where \code{0.n}
indicates a quantile.}
}
