% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_chat_to_code.R
\name{write_chat_to_code}
\alias{write_chat_to_code}
\title{Write Chat Text to Code}
\usage{
write_chat_to_code(filename, folder, chat_response)
}
\arguments{
\item{filename}{a string indicating what the file should be called}

\item{folder}{a string indicating where the file should be stored}

\item{chat_response}{a string from the LLM chat API with code chunks}
}
\value{
Nothing is returned - side effects only to write an R file
}
\description{
This function takes chat responses with code chunks and converts it
to R files.
}
