% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotations.R
\name{annotateCTSS}
\alias{annotateCTSS}
\alias{annotateConsensusClusters}
\title{Annotate and compute summary statistics}
\usage{
annotateCTSS(object, ranges)

annotateConsensusClusters(object, ranges)
}
\arguments{
\item{object}{A \code{\link{CAGEexp}} object (\code{\link{CAGEset}}s are
not supported).}

\item{ranges}{A \code{\link{GRanges}} object containing \code{gene_name},
\code{type} and \code{transcript_type} metadata.}
}
\value{
\code{annotateCTSS} returns the input object with the following modifications:

\itemize{
  \item The Genomic Ranges of the \code{tagCountMatrix} experiment gains a
    \code{annotation} metadata column, with levels such as \dQuote{promoter},
    \dQuote{exon}, \dQuote{intron} and \dQuote{unknown"}, and a \code{genes}
    column, with gene symbols from the annotation.
  \item New column data added, indicating total counts in each of the annotation
    levels.
}

\code{annotateConsensusClusters} returns the input object with the following
modification:

\itemize{
  \item The Genomic Ranges of the \code{consensusClusters} experiment gains a
    \code{annotation} metadata column, with levels such as \dQuote{promoter},
    \dQuote{exon}, \dQuote{intron} and \dQuote{unknown"}, and a \code{genes}
    column, with gene symbols from the annotation.
}
}
\details{
\code{annotateCTSS} annotates the \emph{CTSS} of a CAGEr object and computes
annotation statistics.

\code{annotateConsensusClusters} annotates the \emph{consensus clusters}
of a CAGEr object.
}
\examples{
ce <- readRDS(system.file(package = "CAGEr", "extdata/CAGEexp.rds"))
gff <- readRDS(system.file("extdata/Zv9_annot.rds", package = "CAGEr"))
annotateCTSS(ce, gff)
colData(ce)

normalizeTagCount(ce)
clusterCTSS( object = ce, threshold = 50, thresholdIsTpm = TRUE
           , nrPassThreshold = 1, method = "distclu", maxDist = 20
           , removeSingletons = TRUE, keepSingletonsAbove = 100)
aggregateTagClusters(ce, tpmThreshold = 50, excludeSignalBelowThreshold = FALSE, maxDist = 100)
annotateConsensusClusters(ce, gff)
consensusClustersGR(ce)

}
\seealso{
The \code{\link{Zv9_annot}} example data.

Other CAGEr object modifiers: \code{\link{CTSStoGenes}},
  \code{\link{aggregateTagClusters}},
  \code{\link{clusterCTSS}},
  \code{\link{cumulativeCTSSdistribution}},
  \code{\link{getCTSS}}, \code{\link{normalizeTagCount}},
  \code{\link{quantilePositions}}

Other CAGEr annotation functions: \code{\link{plotAnnot}},
  \code{\link{ranges2annot}}, \code{\link{ranges2genes}}
}
\author{
Charles Plessy
}
