% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotations.R
\docType{methods}
\name{annotateCTSS}
\alias{annotateCTSS}
\alias{annotateCTSS,CAGEset,ANY-method}
\alias{annotateCTSS,CAGEexp,GRanges-method}
\alias{annotateConsensusClusters}
\alias{annotateConsensusClusters,CAGEset,ANY-method}
\alias{annotateConsensusClusters,CAGEexp,GRanges-method}
\title{Annotate and compute summary statistics}
\usage{
annotateCTSS(object, ranges)

\S4method{annotateCTSS}{CAGEset,ANY}(object, ranges)

\S4method{annotateCTSS}{CAGEexp,GRanges}(object, ranges)

annotateConsensusClusters(object, ranges)

\S4method{annotateConsensusClusters}{CAGEset,ANY}(object, ranges)

\S4method{annotateConsensusClusters}{CAGEexp,GRanges}(object, ranges)
}
\arguments{
\item{object}{A \code{\link{CAGEexp}} object (\code{\link{CAGEset}}s are
not supported).}

\item{ranges}{A \code{\link{GRanges}} object, optionally containing \code{gene_name},
\code{type} and \code{transcript_type} metadata.}
}
\value{
\code{annotateCTSS} returns the input object with the following modifications:

\itemize{
\item The Genomic Ranges of the \code{tagCountMatrix} experiment gains a
\code{annotation} metadata column, with levels such as \dQuote{promoter},
\dQuote{exon}, \dQuote{intron} and \dQuote{unknown}.  If the annotation
has a \code{gene_name} metadata, then a \code{genes} column is also added,
with gene symbols from the annotation.
\item The sample data get new columns, indicating total counts in each of the
annotation levels.   If the annotation has a \code{gene_name} metadata, then
a \code{genes} column is added to indicate the number of different gene symbols
detected.
}

\code{annotateConsensusClusters} returns the input object with the same
modifications as above.
}
\description{
\code{annotateCTSS} annotates the \emph{CTSS} of a CAGEr object and computes
annotation statistics.

\code{annotateConsensusClusters} annotates the \emph{consensus clusters}
of a CAGEr object.
}
\examples{
library(SummarizedExperiment)
annotateCTSS(exampleCAGEexp, exampleZv9_annot)
colData(exampleCAGEexp)

annotateConsensusClusters(exampleCAGEexp, exampleZv9_annot)
consensusClustersGR(exampleCAGEexp)

}
\seealso{
The \code{\link{exampleZv9_annot}} example data.

Other CAGEr object modifiers: \code{\link{CTSStoGenes}},
  \code{\link{CustomConsensusClusters}},
  \code{\link{aggregateTagClusters}},
  \code{\link{clusterCTSS}},
  \code{\link{cumulativeCTSSdistribution}},
  \code{\link{getCTSS}}, \code{\link{normalizeTagCount}},
  \code{\link{quantilePositions}},
  \code{\link{summariseChrExpr}}

Other CAGEr annotation functions: \code{\link{plotAnnot}},
  \code{\link{ranges2annot}}, \code{\link{ranges2genes}},
  \code{\link{ranges2names}}
}
\author{
Charles Plessy
}
