% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotations.R
\name{annotateCTSS}
\alias{annotateCTSS}
\title{Annotate and compute summary statistics.}
\usage{
annotateCTSS(object, ranges)
}
\arguments{
\item{object}{A \code{\link{CAGEexp}} object (\code{\link{CAGEset}}s are
not supported).}

\item{ranges}{A \code{\link{GRanges}} object containing \code{gene_name},
\code{type} and \code{transcript_type} metadata.}
}
\value{
The input object with the following modifications:

\itemize{
  \item The Genomic Ranges of the \code{tagCountMatrix} experiment cains a
    \code{annotation} metadata column, with levels such as \dQuote{promoter},
    \dQuote{exon}, \dQuote{intron} and \dQuote{unknown"}.
  \item Column data are added with the raw counts in each of the annotation
    levels.
}
}
\details{
Annotate the CTSS of a CAGEr object and compute annotation statistics.
}
\examples{
\dontrun{
annotateCTSS(object)
}

}
\seealso{
The \code{\link{Zv9_annot}} example data.

Other CAGEr object modifiers: \code{\link{getCTSS}},
  \code{\link{normalizeTagCount}}

Other CAGEr annotation functions: \code{\link{plotAnnot}},
  \code{\link{ranges2annot}}, \code{\link{ranges2genes}}
}
\author{
Charles Plessy
}
