% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportMethods.R
\name{bam2CTSS}
\alias{bam2CTSS}
\title{bam2CTSS}
\usage{
bam2CTSS(gr, removeFirstG, correctSystematicG, genome)
}
\arguments{
\item{gr}{A \code{\link{GRanges}} object returned by \code{\link{import.bam}}.}

\item{removeFirstG}{See getCTSS().}

\item{correctSystematicG}{See getCTSS().}

\item{genome}{See coerceInBSgenome().}
}
\value{
Returns a \code{\link{CTSS}} object.
}
\description{
Converts from BAM to CTSS
}
\details{
Converts genomic ranges representing SAM/BAM alignments into a CTSS object.
}
\seealso{
Other loadFileIntoGPos: 
\code{\link{import.CTSS}()},
\code{\link{import.bam}()},
\code{\link{import.bam.ctss}()},
\code{\link{import.bedCTSS}()},
\code{\link{import.bedScore}()},
\code{\link{import.bedmolecule}()},
\code{\link{import.bigwig}()},
\code{\link{loadFileIntoGPos}()},
\code{\link{moleculesGR2CTSS}()}
}
\concept{loadFileIntoGPos}
