% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAGEexp.R
\docType{class}
\name{CAGEexp-class}
\alias{CAGEexp-class}
\alias{CAGEexp}
\title{CAGEr class to hold all data and metadata about one CAGE experiment.}
\description{
The \code{CAGEr} class is a \code{\link[MultiAssayExperiment]{MultiAssayExperiment}}
object containing all data and metadata about a set of CAGE libraries.  It
is a replacement for the \code{\link{CAGEset}} class.  The main difference
is that the expression data is stored in \code{\link[S4Vectors]{DataFrame}} objects
of \code{\link[S4Vectors]{Rle}}-encoded expression values, instead of plain
\code{data.frame}s.  With large datasets, this saves considerable amounts of memory.
}
\details{
If \code{genomeName} is \code{NULL}, checks of chromosome names will be
disabled and G-correction will not be possible.  See https://support.bioconductor.org/p/86437/
for an example on how to create a BSgenome package.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{A list that must at least contain \code{genomeName} and
\code{inputFilesType} members.}
}}

\examples{

library("MultiAssayExperiment")
library("SummarizedExperiment")
pathsToInputFiles <- list.files( system.file("extdata", package = "CAGEr")
                               , "ctss$"
                               , full.names = TRUE)
sampleLabels <- sub( ".chr17.ctss", "", basename(pathsToInputFiles))
ce <- CAGEexp( metadata = list(genomeName = "BSgenome.Drerio.UCSC.danRer7")
             , colData  = DataFrame( inputFiles     = pathsToInputFiles
                                   , sampleLabels   = sampleLabels
                                   , inputFilesType = "ctss"
                                   , row.names      = sampleLabels))

# Expression data is loaded by the getCTSS() function, that also calculates
# library sizes and store them in the object's column data.

getCTSS(ce)
librarySizes(ce)
colData(ce)

# CTSS data is stored internally as a SummarizedExperiemnt that can be retreived
# as a whole, or as GRanges, or as an expression DataFrame.

CTSStagCountSE(ce)
CTSScoordinatesGR(ce)
CTSStagCountDF(ce)

# Columns of the "colData" table are accessible directly via the "$" operator.

ce$l1 <- colSums(CTSStagCountDf(ce) > 0)
ce$l1
 
# The commands below were used to create the example CAGEexp object in the "extdata" folder.
\dontrun{
normalizeTagCount(ce)
clusterCTSS(ce)
cumulativeCTSSdistribution(ce, "tagClusters")
quantilePositions(ce, "tagClusters")
aggregateTagClusters(ce)
cumulativeCTSSdistribution(ce, "consensusClusters")
quantilePositions(ce, "consensusClusters")
saveRDS(ce, file = "inst/extdata/CAGEexp.rds", compress = "xz")
}

}
\seealso{
CAGEset-class
}
