% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAGEexp.R
\docType{class}
\name{CAGEexp-class}
\alias{CAGEexp-class}
\alias{CAGEexp}
\alias{coerce}
\title{CAGEr class to hold all data and metadata about one CAGE experiment.}
\description{
The \code{CAGEr} class is a \code{\link[MultiAssayExperiment]{MultiAssayExperiment}}
object containing all data and metadata about a set of CAGE libraries.  It
is a replacement for the \code{\link{CAGEset}} class.  The main difference
is that the expression data is stored in \code{\link[S4Vectors]{DataFrame}} objects
of \code{\link[S4Vectors]{Rle}}-encoded expression values, instead of plain
\code{data.frame}s.  With large datasets, this saves considerable amounts of memory.
}
\details{
If \code{genomeName} is \code{NULL}, checks of chromosome names will be
disabled and G-correction will not be possible.  See https://support.bioconductor.org/p/86437/
for an example on how to create a BSgenome package.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{A list that must at least contain \code{genomeName} and
\code{inputFilesType} members.}
}}

\examples{
library("MultiAssayExperiment")
library("SummarizedExperiment")
pathsToInputFiles <- list.files( system.file("extdata", package = "CAGEr")
                               , "ctss$"
                               , full.names = TRUE)
sampleLabels <- sub( ".chr17.ctss", "", basename(pathsToInputFiles))
exampleCAGEexp <-
  CAGEexp( metadata = list(genomeName = "BSgenome.Drerio.UCSC.danRer7")
         , colData  = DataFrame( inputFiles     = pathsToInputFiles
                               , sampleLabels   = sampleLabels
                               , inputFilesType = "ctss"
                               , row.names      = sampleLabels))

# Expression data is loaded by the getCTSS() function, that also calculates
# library sizes and store them in the object's column data.

getCTSS(exampleCAGEexp)
librarySizes(exampleCAGEexp)
colData(exampleCAGEexp)

# CTSS data is stored internally as a SummarizedExperiemnt that can be retreived
# as a whole, or as GRanges, or as an expression DataFrame.

CTSStagCountSE(exampleCAGEexp)
CTSScoordinatesGR(exampleCAGEexp)
CTSStagCountDF(exampleCAGEexp)

# Columns of the "colData" table are accessible directly via the "$" operator.

exampleCAGEexp$l1 <- colSums(CTSStagCountDf(exampleCAGEexp) > 0)
exampleCAGEexp$l1
 
# The commands below were used to create the example CAGEexp object in the "extdata" folder.
\dontrun{
exampleCAGEexp <- exampleCAGEexp[,c(5, 2, 1, 3, 4)] # Non-aplhabetic order may help catch bugs
CTSStagCountSE(exampleCAGEexp) <- CTSStagCountSE(exampleCAGEexp)[1:5000,]  # Slim the object
annotateCTSS(exampleCAGEexp , exampleZv9_annot)
CTSStoGenes(exampleCAGEexp)
normalizeTagCount(exampleCAGEexp)
clusterCTSS(exampleCAGEexp)
cumulativeCTSSdistribution(exampleCAGEexp, "tagClusters")
quantilePositions(exampleCAGEexp, "tagClusters")
aggregateTagClusters(exampleCAGEexp)
cumulativeCTSSdistribution(exampleCAGEexp, "consensusClusters")
quantilePositions(exampleCAGEexp, "consensusClusters")
save(exampleCAGEexp, file = "data/exampleCAGEexp.RData", compress = "xz")
}

}
\seealso{
CAGEset-class
}
