% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringMethods.R
\name{.clusterAggregateAndSum}
\alias{.clusterAggregateAndSum}
\alias{.clusterAggregateAndSum,data.table-method}
\alias{.clusterAggregateAndSum,data.frame-method}
\alias{.clusterAggregateAndSum,GRanges-method}
\title{Aggregate identical clusters and sum their scores.}
\usage{
.clusterAggregateAndSum(clusters, key)

\S4method{.clusterAggregateAndSum}{data.table}(clusters, key)

\S4method{.clusterAggregateAndSum}{data.frame}(clusters, key)

\S4method{.clusterAggregateAndSum}{GRanges}(clusters, key)
}
\arguments{
\item{clusters}{Clusters to be aggregated.  \code{data.frame}, or
\code{GRanges}, which will be coerced to \code{data.frame}.}

\item{key}{Name of the column containing the factor used to aggregate
the clusters.}
}
\description{
Private function using \code{data.table} objects to preform grouping
operations at a high performance.  These functions use \emph{non-standard
evaluation} in a context that raises warnings in \verb{R CMD check}.  By
separating these functions from the rest of the code, I hope to make the
workarounds easier to manage.
}
