% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{consensusClusterConvertors}
\alias{consensusClusterConvertors}
\alias{CCgranges2dataframe}
\alias{CCdataframe2granges}
\title{Private functions to convert CC formats}
\usage{
CCgranges2dataframe(gr)

CCdataframe2granges(df)
}
\arguments{
\item{gr}{Consensus clusters in \code{GRanges} format.}

\item{df}{Consensus clusters in \code{data.frame} format.}
}
\description{
Interconvert consensus clusters (CC) formats used in classes CAGEset
(\code{data.frame}) and CAGEexp (\code{GRanges}).
}
\examples{
df <- consensusClusters(exampleCAGEset)
head(df)
gr <- CCdataframe2granges(df)
gr
# No round-trip because start and end were not integer in df.
# identical(df, CCgranges2dataframe(gr))
}
\seealso{
Other df2granges converters: \code{\link{tagClusterConvertors}}
}
\concept{df2granges converters}
