% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTSS.R
\docType{class}
\name{ConsensusClusters}
\alias{ConsensusClusters}
\alias{.ConsensusClusters}
\title{Consensus clusters class}
\details{
The \code{ConsensusClusters} class represents consensus clusters.
It is used internally by CAGEr for type safety.
}
