% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R, R/SetMethods.R
\docType{methods}
\name{consensusClusters}
\alias{consensusClusters}
\alias{consensusClusters,CAGEr-method}
\alias{consensusClusters<-}
\alias{consensusClusters<-,CAGEset-method}
\alias{consensusClusters<-,CAGEexp-method}
\alias{consensusClustersGR<-}
\alias{consensusClustersGR<-,CAGEset-method}
\alias{consensusClustersGR<-,CAGEexp-method}
\alias{consensusClustersSE<-}
\alias{consensusClustersSE<-,CAGEset,ANY-method}
\alias{consensusClustersSE<-,CAGEexp,RangedSummarizedExperiment-method}
\title{Get or set consensus clusters from CAGEr objects

Extracts or inserts the information on consensus clusters from a CAGEr object.}
\usage{
consensusClusters(object, sample = NULL, returnInterquantileWidth = FALSE,
  qLow = NULL, qUp = NULL)

\S4method{consensusClusters}{CAGEr}(object, sample = NULL,
  returnInterquantileWidth = FALSE, qLow = NULL, qUp = NULL)

consensusClusters(object) <- value

\S4method{consensusClusters}{CAGEset}(object) <- value

\S4method{consensusClusters}{CAGEexp}(object) <- value

consensusClustersGR(object) <- value

\S4method{consensusClustersGR}{CAGEset}(object) <- value

\S4method{consensusClustersGR}{CAGEexp}(object) <- value

consensusClustersSE(object) <- value

\S4method{consensusClustersSE}{CAGEset,ANY}(object) <- value


  \S4method{consensusClustersSE}{CAGEexp,RangedSummarizedExperiment}(object) <- value
}
\arguments{
\item{object}{A \code{\link{CAGEr}} object.}

\item{sample}{Optional. Label of the CAGE dataset (experiment, sample) for which to extract
sample-specific information on consensus clusters.}

\item{returnInterquantileWidth}{Should the interquantile width of consensus clusters in
specified sample be returned.  Used only when \code{sample} argument is specified, otherwise
ignored.}

\item{qLow}{Position of which quantile should be used as a left (lower) boundary when
calculating interquantile width.  Used only when \code{sample} argument is specified and
\code{returnInterquantileWidth = TRUE}, otherwise ignored.}

\item{qUp}{Position of which quantile should be used as a right (upper) boundary when
calculating interquantile width.  Used only when \code{sample} argument is specified and
\code{returnInterquantileWidth = TRUE}, otherwise ignored.}

\item{value}{A \code{data.frame} of consensus clusters}
}
\value{
Returns a \code{data.frame} with information on consensus clusters, including genomic
coordinates.  When \code{sample} argument is NOT specified, total CAGE signal across all CAGE
datasets (samples) is returned in the \code{tpm} column.  When \code{sample} argument is
specified, the \code{tpm} column contains CAGE signal of consensus clusters in that specific
sample.  When \code{returnInterquantileWidth = TRUE}, additional sample-specific information
is returned, including position of the dominant TSS, and interquantile width of the consensus
clusters in the specified sample.
}
\examples{
load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
clusters.general <- consensusClusters(exampleCAGEset)
head(clusters.general)

clusters.sample <- consensusClusters(exampleCAGEset, sample = "sample2")
head(clusters.sample)

}
\seealso{
Other CAGEr clusters functions: \code{\link{CTSScumulativesTagClusters}},
  \code{\link{aggregateTagClusters}},
  \code{\link{clusterCTSS}},
  \code{\link{consensusClustersDESeq2}},
  \code{\link{cumulativeCTSSdistribution}},
  \code{\link{plotInterquantileWidth}},
  \code{\link{quantilePositions}},
  \code{\link{tagClusters}}
}
\author{
Vanja Haberle
}
