% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\name{consensusClustersGR}
\alias{consensusClustersGR}
\alias{consensusClustersGR,CAGEexp-method}
\alias{consensusClustersSE}
\alias{consensusClustersSE,CAGEexp-method}
\title{Get consensus clusters from CAGEr objects}
\usage{
consensusClustersGR(
  object,
  sample = NULL,
  returnInterquantileWidth = FALSE,
  qLow = NULL,
  qUp = NULL
)

\S4method{consensusClustersGR}{CAGEexp}(
  object,
  sample = NULL,
  returnInterquantileWidth = FALSE,
  qLow = NULL,
  qUp = NULL
)

consensusClustersSE(object)

\S4method{consensusClustersSE}{CAGEexp}(object)
}
\arguments{
\item{object}{A \code{\link{CAGEr}} object.}

\item{sample}{Optional. Label of the CAGE dataset (experiment, sample) for
which to extract sample-specific information on consensus clusters.}

\item{returnInterquantileWidth}{Should the interquantile width of consensus
clusters in specified sample be returned.  Used only when \code{sample}
argument is specified, otherwise ignored.}

\item{qLow, qUp}{Position of which quantile should be used as a left (lower)
or right (upper) boundary when calculating interquantile width.  Used
only when \code{sample} argument is specified and
\code{returnInterquantileWidth = TRUE}, otherwise ignored.}
}
\value{
\code{consensusClustersGR} returns a \code{\link{ConsensusClusters}} object, which
wraps the \code{\link{GRanges}} class.  The \code{score} columns indicates the
normalised expression value of each cluster, either across all samples
(\code{sample = NULL}), or for the selected sample.  The legacy \code{tpm} column may
be removed in the future.  When \code{sample} argument is
NOT specified, total CAGE signal across all CAGE datasets (samples) is
returned in the \code{tpm} column.  When \code{sample} argument is specified, the \code{tpm}
column contains CAGE signal of consensus clusters in that specific sample.
When \code{returnInterquantileWidth = TRUE}, additional sample-specific information
is returned, including position of the dominant TSS, and interquantile width
of the consensus clusters in the specified sample.

\code{consensusClustersSE} returns the \code{\link{SummarizedExperiment}} stored
in the \code{consensusClusters} experiment slot of the CAGEexp object.
}
\description{
Extracts the information on consensus clusters from a \code{\link{CAGEr}}
object.
}
\examples{
consensusClustersGR( exampleCAGEexp, sample = 2
                   , returnInterquantileWidth = TRUE
                   , qLow = 0.1, qUp = 0.9)

}
\seealso{
\code{\link[=consensusClusters<-]{consensusClusters<-()}}

Other CAGEr accessor methods: 
\code{\link{CTSSclusteringMethod}()},
\code{\link{CTSScoordinatesGR}()},
\code{\link{CTSScumulativesTagClusters}()},
\code{\link{CTSSnormalizedTpmDF}()},
\code{\link{CTSStagCountDF}()},
\code{\link{GeneExpDESeq2}()},
\code{\link{GeneExpSE}()},
\code{\link{expressionClasses}()},
\code{\link{genomeName}()},
\code{\link{inputFilesType}()},
\code{\link{inputFiles}()},
\code{\link{librarySizes}()},
\code{\link{sampleLabels}()},
\code{\link{seqNameTotalsSE}()},
\code{\link{tagClustersGR}()}

Other CAGEr clusters functions: 
\code{\link{CTSSclusteringMethod}()},
\code{\link{CTSScumulativesTagClusters}()},
\code{\link{CustomConsensusClusters}()},
\code{\link{aggregateTagClusters}()},
\code{\link{clusterCTSS}()},
\code{\link{consensusClustersDESeq2}()},
\code{\link{cumulativeCTSSdistribution}()},
\code{\link{plotInterquantileWidth}()},
\code{\link{quantilePositions}()},
\code{\link{tagClustersGR}()}
}
\author{
Vanja Haberle

Charles Plessy
}
\concept{CAGEr accessor methods}
\concept{CAGEr clusters functions}
