% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\docType{methods}
\name{consensusClustersGR}
\alias{consensusClustersGR}
\alias{consensusClustersGR,CAGEset-method}
\alias{consensusClustersGR,CAGEexp-method}
\alias{consensusClustersSE}
\alias{consensusClustersSE,CAGEset-method}
\alias{consensusClustersSE,CAGEexp-method}
\alias{consensusClusters}
\alias{consensusClusters,CAGEr-method}
\title{Get consensus clusters from CAGEr objects}
\usage{
consensusClustersGR(object, sample = NULL,
  returnInterquantileWidth = FALSE, qLow = NULL, qUp = NULL)

\S4method{consensusClustersGR}{CAGEset}(object, sample = NULL,
  returnInterquantileWidth = FALSE, qLow = NULL, qUp = NULL)

\S4method{consensusClustersGR}{CAGEexp}(object, sample = NULL,
  returnInterquantileWidth = FALSE, qLow = NULL, qUp = NULL)

consensusClustersSE(object)

\S4method{consensusClustersSE}{CAGEset}(object)

\S4method{consensusClustersSE}{CAGEexp}(object)

consensusClusters(object, sample = NULL,
  returnInterquantileWidth = FALSE, qLow = NULL, qUp = NULL)

\S4method{consensusClusters}{CAGEr}(object, sample = NULL,
  returnInterquantileWidth = FALSE, qLow = NULL, qUp = NULL)
}
\arguments{
\item{object}{A \code{\link{CAGEr}} object.}

\item{sample}{Optional. Label of the CAGE dataset (experiment, sample) for
which to extract sample-specific information on consensus clusters.}

\item{returnInterquantileWidth}{Should the interquantile width of consensus
clusters in specified sample be returned.  Used only when \code{sample}
argument is specified, otherwise ignored.}

\item{qLow, qUp}{Position of which quantile should be used as a left (lower)
or right (upper) boundary when calculating interquantile width.  Used
only when \code{sample} argument is specified and
\code{returnInterquantileWidth = TRUE}, otherwise ignored.}
}
\value{
\code{consensusClustersGR} returns a \code{\link{ConsensusClusters}} object, which
is a \code{\link{GRanges}} wrapper containing similar information as the data frame
returned by \code{consensusClustersGR}.  The \code{score} columns indicates the
normalised expression value of each cluster, either across all samples
(\code{sample = NULL}), or for the selected sample.  The \code{tpm} column provides
the same information for compatibility with \code{CAGEset} objects but may be
removed in the future.

\code{consensusClustersSE} returns the \code{\link{SummarizedExperiment}} stored
in the \code{consensusClusters} experiment slot of the CAGEexp object.

\code{consensusClusters} returns a \code{data.frame} with information on
consensus clusters, including genomic coordinates.  When \code{sample} argument is
NOT specified, total CAGE signal across all CAGE datasets (samples) is
returned in the \code{tpm} column.  When \code{sample} argument is specified, the \code{tpm}
column contains CAGE signal of consensus clusters in that specific sample.
When \code{returnInterquantileWidth = TRUE}, additional sample-specific information
is returned, including position of the dominant TSS, and interquantile width
of the consensus clusters in the specified sample.
}
\description{
Extracts the information on consensus clusters from a \code{\link{CAGEr}}
object.
}
\examples{
consensusClustersGR( exampleCAGEexp, sample = 2
                   , returnInterquantileWidth = TRUE
                   , qLow = 0.1, qUp = 0.9)
head(consensusClusters(exampleCAGEset))
head(consensusClusters(exampleCAGEset, sample = "sample2"))

cumulativeCTSSdistribution(exampleCAGEset, "consensusClusters") # Defaults in object do not fit
quantilePositions(exampleCAGEset, "consensusClusters")
head(consensusClusters(exampleCAGEset, sample = "sample2"
                      , returnInterquantileWidth = TRUE
                      , qLow = 0.1, qUp = 0.9))

head(consensusClusters(exampleCAGEexp))
consensusClustersGR(exampleCAGEexp, 2)

}
\seealso{
\code{\link[=consensusClusters<-]{consensusClusters<-()}}

Other CAGEr accessor methods: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScoordinates}},
  \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CTSSnormalizedTpm}},
  \code{\link{CTSStagCountTable}},
  \code{\link{CTSStagCount}}, \code{\link{GeneExpDESeq2}},
  \code{\link{GeneExpSE}}, \code{\link{genomeName}},
  \code{\link{inputFilesType}}, \code{\link{inputFiles}},
  \code{\link{librarySizes}}, \code{\link{sampleLabels}},
  \code{\link{seqNameTotalsSE}}, \code{\link{tagClusters}}

Other CAGEr clusters functions: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CustomConsensusClusters}},
  \code{\link{aggregateTagClusters}},
  \code{\link{clusterCTSS}},
  \code{\link{consensusClustersDESeq2}},
  \code{\link{cumulativeCTSSdistribution}},
  \code{\link{plotInterquantileWidth}},
  \code{\link{quantilePositions}},
  \code{\link{tagClusters}}
}
\author{
Vanja Haberle
}
\concept{CAGEr accessor methods}
\concept{CAGEr clusters functions}
