% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R, R/SetMethods.R
\docType{methods}
\name{consensusClustersQuantile}
\alias{consensusClustersQuantile}
\alias{consensusClustersQuantileLow}
\alias{consensusClustersQuantileLow,CAGEset-method}
\alias{consensusClustersQuantileLow,CAGEexp-method}
\alias{consensusClustersQuantileUp}
\alias{consensusClustersQuantileUp,CAGEset-method}
\alias{consensusClustersQuantileUp,CAGEexp-method}
\alias{consensusClustersQuantile,CAGEset-method}
\alias{consensusClustersQuantile,CAGEexp-method}
\alias{consensusClustersQuantileLow<-}
\alias{consensusClustersQuantileLow<-,CAGEset-method}
\alias{consensusClustersQuantileUp<-}
\alias{consensusClustersQuantileUp<-,CAGEset-method}
\title{Quantile metadata stored in CAGEr objects.}
\usage{
consensusClustersQuantileLow(object, samples = NULL)

\S4method{consensusClustersQuantileLow}{CAGEset}(object, samples = NULL)

\S4method{consensusClustersQuantileLow}{CAGEexp}(object, samples = NULL)

consensusClustersQuantileUp(object, samples = NULL)

\S4method{consensusClustersQuantileUp}{CAGEset}(object, samples = NULL)

\S4method{consensusClustersQuantileUp}{CAGEexp}(object, samples = NULL)

consensusClustersQuantile(object, sample = NULL, q)

\S4method{consensusClustersQuantile}{CAGEset}(object, sample = NULL, q)

\S4method{consensusClustersQuantile}{CAGEexp}(object, sample = NULL, q)

consensusClustersQuantileLow(object, samples = NULL) <- value

\S4method{consensusClustersQuantileLow}{CAGEset}(object,
  samples = NULL) <- value

consensusClustersQuantileUp(object, samples = NULL) <- value

\S4method{consensusClustersQuantileUp}{CAGEset}(object,
  samples = NULL) <- value
}
\arguments{
\item{object}{A \code{\link{CAGEr}} object.}

\item{samples}{Sample name(s), number(s) or \code{NULL} (default) for all samples.}

\item{sample}{A single sample name or number, or \code{NULL} (default) for all samples.}

\item{q}{A quantile.}

\item{value}{A list (one entry per sample) of data frames with multiple columns:
\code{cluster} for the cluster ID, and then \code{q_0.n} where \code{0.n}
indicates a quantile.}
}
\description{
Accessors for consensus cluster quantile data in CAGEr objects.
}
