% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\docType{methods}
\name{consensusClustersTpm}
\alias{consensusClustersTpm}
\alias{consensusClustersTpm,CAGEset-method}
\alias{consensusClustersTpm,CAGEexp-method}
\alias{consensusClustersTpmDf}
\alias{consensusClustersTpmDf,CAGEset-method}
\alias{consensusClustersTpmDf,CAGEexp-method}
\alias{consensusClustersTpmDF}
\alias{consensusClustersTpmDF,CAGEset-method}
\alias{consensusClustersTpmDF,CAGEexp-method}
\title{Extracting consensus clusters tpm matrix from CAGEset object}
\usage{
consensusClustersTpm(object)
consensusClustersTpmDf(object)
consensusClustersTpmDF(object)

\S4method{consensusClustersTpm}{CAGEset}(object)

\S4method{consensusClustersTpm}{CAGEexp}(object)

consensusClustersTpmDf(object)

\S4method{consensusClustersTpmDf}{CAGEset}(object)

\S4method{consensusClustersTpmDf}{CAGEexp}(object)

consensusClustersTpmDF(object)

\S4method{consensusClustersTpmDF}{CAGEset}(object)

\S4method{consensusClustersTpmDF}{CAGEexp}(object)
}
\arguments{
\item{object}{A CAGEr object.}
}
\value{
Returns normalized expression valuse of CAGE clusters across all samples.
\itemize{
  \item For \code{consensusClustersTpm}, a \code{matrix}.
  \item For \code{consensusClustersTpmDf}, a \code{data.frame}.
  \item For \code{consensusClustersTpmDF}, a \code{DataFrame}.
}
}
\description{
Extracts a table with normalized CAGE tag values for consensus
clusters across all samples from a \code{\link{CAGEr}} object.
}
\examples{
load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
clusters.tpm <- consensusClustersTpm(exampleCAGEset)
head(clusters.tpm)
consensusClustersTpmDF(exampleCAGEset)

}
\seealso{
\code{\link{consensusClusters}}
}
\author{
Vanja Haberle
}
