% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTSS.R
\docType{class}
\name{CTSS-class}
\alias{CTSS-class}
\alias{.CTSS}
\alias{CTSS.chr-class}
\alias{.CTSS.chr}
\title{CAGE Transcription Start Site class}
\details{
The \code{CTSS} class represents CAGE transcription start sites (CTSS) at
single-nucleotide resolution, using GRanges as base class.  It is used internally by CAGEr
for type safety.

The \code{CTSS.chr} class represents a \code{CTSS} object that is guaranteed to be
only on a single chromosome.  It is used internally by CAGEr for type safety.
}
\examples{
ce <- readRDS(system.file(package = "CAGEr", "extdata/CAGEexp.rds"))
ctss <- CAGEr:::.CTSS(CTSScoordinatesGR(ce))
ce <- readRDS(system.file(package = "CAGEr", "extdata/CAGEexp.rds"))
ctss.chr <- CAGEr:::.CTSS.chr(CTSScoordinatesGR(ce))
}
\author{
Charles Plessy
}
