% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTSS.R
\docType{class}
\name{CTSS-class}
\alias{CTSS-class}
\alias{.CTSS}
\alias{CTSS}
\alias{CTSS.chr-class}
\alias{.CTSS.chr}
\alias{CTSS.chr}
\alias{ConsensusClusters}
\alias{.ConsensusClusters}
\alias{TagClusters}
\alias{.TagClusters}
\title{Classes for type safety.}
\description{
The \code{CTSS} class represents CAGE transcription start sites
(CTSS) at single-nucleotide resolution, using \code{GPos} as base class.  It is
used internally by \emph{CAGEr} for type safety.

The \code{ConsensusClusters} class represents consensus clusters.
It is used internally by CAGEr for type safety.
}
\details{
The \code{genomeName} element of the \code{metadata} slot is used to store the
name of the \emph{BSgenome} package used when constructing the \code{CAGEr} object.
Be careful that calling the \code{CTSS} constructor twice in a row will erase
the fist metadata (because the \code{\link{GPos}} constructor does so).  \emph{CAGEr}
uses the \code{UnstitchedGPos} implementation of \code{GPos}.

The \code{CTSS.chr} class represents a \link{CTSS} object that is guaranteed
to be only on a single chromosome.  It is used internally by \emph{CAGEr} for type
safety.

Consensus clusters must not overlap, so that a single TSS in the
genome can only be attributed to a single cluster.

The \code{TagClusters} class represents tag clusters.
It is used internally by CAGEr for type safety.
}
\examples{
ctss <- CAGEr:::.CTSS( CTSScoordinatesGR(exampleCAGEexp)
                     , bsgenomeName = genomeName(exampleCAGEexp))
genomeName(ctss)

ctss.chr <- as(CTSScoordinatesGR(exampleCAGEexp), "CTSS.chr")
}
\author{
Charles Plessy
}
