% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R, R/SetMethods.R
\docType{methods}
\name{CTSSclusteringMethod}
\alias{CTSSclusteringMethod}
\alias{CTSSclusteringMethod,CAGEset-method}
\alias{CTSSclusteringMethod,CAGEexp-method}
\alias{CTSSclusteringMethod<-}
\alias{CTSSclusteringMethod<-,CAGEset-method}
\alias{CTSSclusteringMethod<-,CAGEexp-method}
\title{Extracting CTSS clustering method from CAGEr objects.}
\usage{
CTSSclusteringMethod(object)

\S4method{CTSSclusteringMethod}{CAGEset}(object)

\S4method{CTSSclusteringMethod}{CAGEexp}(object)

CTSSclusteringMethod(object) <- value

\S4method{CTSSclusteringMethod}{CAGEset}(object) <- value

\S4method{CTSSclusteringMethod}{CAGEexp}(object) <- value
}
\arguments{
\item{object}{A CAGEr object.}

\item{value}{character}
}
\value{
Returns a label of the method used for CTSS clustering.
}
\description{
Extracts the label of the method used for CTSS clustering into tag
clusters from \code{\link{CAGEr}} objects.
}
\examples{
CTSSclusteringMethod(exampleCAGEset)

}
\seealso{
\code{\link{clusterCTSS}}

Other CAGEr accessor methods: \code{\link{CTSScoordinates}},
  \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CTSSnormalizedTpm}},
  \code{\link{CTSStagCountTable}},
  \code{\link{CTSStagCount}}, \code{\link{GeneExpDESeq2}},
  \code{\link{GeneExpSE}}, \code{\link{genomeName}},
  \code{\link{inputFilesType}}, \code{\link{inputFiles}},
  \code{\link{librarySizes}}, \code{\link{sampleLabels}},
  \code{\link{seqNameTotalsSE}}, \code{\link{tagClusters}}
}
\author{
Vanja Haberle
}
