% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R, R/SetMethods.R
\docType{methods}
\name{CTSSclusteringMethod}
\alias{CTSSclusteringMethod}
\alias{CTSSclusteringMethod,CAGEset-method}
\alias{CTSSclusteringMethod,GRangesList-method}
\alias{CTSSclusteringMethod,CAGEexp-method}
\alias{CTSSclusteringMethod<-}
\alias{CTSSclusteringMethod<-,CAGEset-method}
\alias{CTSSclusteringMethod<-,GRangesList-method}
\alias{CTSSclusteringMethod<-,CAGEexp-method}
\title{Get /set CTSS clustering method}
\usage{
CTSSclusteringMethod(object)

\S4method{CTSSclusteringMethod}{CAGEset}(object)

\S4method{CTSSclusteringMethod}{GRangesList}(object)

\S4method{CTSSclusteringMethod}{CAGEexp}(object)

CTSSclusteringMethod(object) <- value

\S4method{CTSSclusteringMethod}{CAGEset}(object) <- value

\S4method{CTSSclusteringMethod}{GRangesList}(object) <- value

\S4method{CTSSclusteringMethod}{CAGEexp}(object) <- value
}
\arguments{
\item{object}{A CAGEr object.}

\item{value}{character}
}
\description{
Returns or sets the name of the method that was used make tag
clusters from the CTSSs of a \code{\link{CAGEr}} object.
}
\examples{
CTSSclusteringMethod(exampleCAGEset)
CTSSclusteringMethod(exampleCAGEexp)

}
\seealso{
\code{\link{clusterCTSS}}

Other CAGEr accessor methods: \code{\link{CTSScoordinates}},
  \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CTSSnormalizedTpm}},
  \code{\link{CTSStagCountTable}},
  \code{\link{CTSStagCount}}, \code{\link{GeneExpDESeq2}},
  \code{\link{GeneExpSE}},
  \code{\link{consensusClustersGR}},
  \code{\link{genomeName}}, \code{\link{inputFilesType}},
  \code{\link{inputFiles}}, \code{\link{librarySizes}},
  \code{\link{sampleLabels}},
  \code{\link{seqNameTotalsSE}}, \code{\link{tagClusters}}

Other CAGEr clusters functions: \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CustomConsensusClusters}},
  \code{\link{aggregateTagClusters}},
  \code{\link{clusterCTSS}},
  \code{\link{consensusClustersDESeq2}},
  \code{\link{consensusClustersGR}},
  \code{\link{cumulativeCTSSdistribution}},
  \code{\link{plotInterquantileWidth}},
  \code{\link{quantilePositions}},
  \code{\link{tagClusters}}
}
\author{
Vanja Haberle

Charles Plessy
}
\concept{CAGEr accessor methods}
\concept{CAGEr clusters functions}
