% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R, R/SetMethods.R
\docType{methods}
\name{CTSScoordinates}
\alias{CTSScoordinates}
\alias{CTSScoordinates,CAGEset-method}
\alias{CTSScoordinates,CAGEexp-method}
\alias{CTSScoordinatesGR}
\alias{CTSScoordinatesGR,CAGEset-method}
\alias{CTSScoordinatesGR,CAGEexp-method}
\alias{CTSScoordinatesGR}
\alias{CTSScoordinatesGR<-,CAGEset-method}
\alias{CTSScoordinatesGR<-,CAGEexp-method}
\alias{CTSStagCountSE}
\alias{CTSStagCountSE<-,CAGEset-method}
\alias{CTSStagCountSE<-,CAGEexp-method}
\title{Extracting genomic coordinates of TSSs from CAGEr object}
\usage{
CTSScoordinates(object)

\S4method{CTSScoordinates}{CAGEset}(object)

\S4method{CTSScoordinates}{CAGEexp}(object)

CTSScoordinatesGR(object)

\S4method{CTSScoordinatesGR}{CAGEset}(object)

\S4method{CTSScoordinatesGR}{CAGEexp}(object)

CTSScoordinatesGR(object) <- value

\S4method{CTSScoordinatesGR}{CAGEset}(object) <- value

\S4method{CTSScoordinatesGR}{CAGEexp}(object) <- value

CTSStagCountSE(object) <- value

\S4method{CTSStagCountSE}{CAGEset}(object) <- value

\S4method{CTSStagCountSE}{CAGEexp}(object) <- value
}
\arguments{
\item{object}{A CAGEset or CAGEexp object.}
}
\value{
\code{CTSScoordinates} returns a \code{data.frame} with genomic coordinates of all
TSSs. \code{pos} column contains 1-based coordinate of the TSS.

\code{CTSScoordinatesGR} returns the coordinates as genomic ranges.  A
\code{filteredCTSSidx} column metadata will be present if \code{\link{clusterCTSS}}
was ran earlier.
}
\description{
Extracts the genomic coordinates of all detected TSSs
from \code{\link{CAGEset}} and \code{\link{CAGEexp}} objects.
}
\examples{
load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
CTSS <- CTSScoordinates(exampleCAGEset)
head(CTSS)

CTSScoordinatesGR(exampleCAGEset)

}
\seealso{
\code{\link{getCTSS}}

\code{\link{clusterCTSS}}

Other CAGEr accessor methods: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CTSSnormalizedTpm}},
  \code{\link{CTSStagCountTable}},
  \code{\link{CTSStagCount}}, \code{\link{genomeName}},
  \code{\link{inputFilesType}}, \code{\link{inputFiles}},
  \code{\link{librarySizes}}, \code{\link{sampleLabels}},
  \code{\link{tagClusters}}
}
\author{
Vanja Haberle

Charles Plessy
}
