% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\docType{methods}
\name{CTSScoordinates}
\alias{CTSScoordinates}
\title{Extracting genomic coordinates of TSSs from CAGEr object}
\usage{
CTSScoordinates(object)
}
\arguments{
\item{object}{A CAGEset or CAGEexp object.}
}
\value{
Returns a \code{data.frame} with genomic coordinates of all TSSs. \code{pos}
column contains 1-based coordinate of the TSS.
}
\description{
Extracts the genomic coordinates of all detected TSSs
from \code{\link{CAGEset}} and \code{\link{CAGEexp}} objects.
}
\examples{
load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
CTSS <- CTSScoordinates(exampleCAGEset)
head(CTSS)

}
\seealso{
\code{\link{getCTSS}}

Other CAGEr accessor methods: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSSnormalizedTpm}},
  \code{\link{CTSStagCountTable}},
  \code{\link{CTSStagCount}}, \code{\link{genomeName}},
  \code{\link{inputFilesType}}, \code{\link{inputFiles}},
  \code{\link{librarySizes}}, \code{\link{sampleLabels}}
}
\author{
Vanja Haberle
}
