% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R, R/SetMethods.R
\docType{methods}
\name{CTSScoordinates}
\alias{CTSScoordinates}
\alias{CTSScoordinates,CAGEset-method}
\alias{CTSScoordinates,CAGEexp-method}
\alias{CTSScoordinatesGR}
\alias{CTSScoordinatesGR,CAGEset-method}
\alias{CTSScoordinatesGR,CAGEexp-method}
\alias{CTSScoordinatesGR<-}
\alias{CTSScoordinatesGR<-,CAGEset-method}
\alias{CTSScoordinatesGR<-,CAGEexp-method}
\alias{CTSStagCountSE<-}
\alias{CTSStagCountSE<-,CAGEset-method}
\alias{CTSStagCountSE<-,CAGEexp-method}
\title{Genomic coordinates of TSSs from a \code{CAGEr} object}
\usage{
CTSScoordinates(object)

\S4method{CTSScoordinates}{CAGEset}(object)

\S4method{CTSScoordinates}{CAGEexp}(object)

CTSScoordinatesGR(object)

\S4method{CTSScoordinatesGR}{CAGEset}(object)

\S4method{CTSScoordinatesGR}{CAGEexp}(object)

CTSScoordinatesGR(object) <- value

\S4method{CTSScoordinatesGR}{CAGEset}(object) <- value

\S4method{CTSScoordinatesGR}{CAGEexp}(object) <- value

CTSStagCountSE(object) <- value

\S4method{CTSStagCountSE}{CAGEset}(object) <- value

\S4method{CTSStagCountSE}{CAGEexp}(object) <- value
}
\arguments{
\item{object}{A \code{CAGEset} or \code{CAGEexp} object.}

\item{value}{Coordinates to update, in a format according to the function name.}
}
\value{
\code{\link[=CTSScoordinates]{CTSScoordinates()}} returns a \code{data.frame} with genomic coordinates
of all TSSs.  The \code{pos} column contains 1-based coordinate of the TSS.

\code{CTSScoordinatesGR} returns the coordinates as a \code{\link[=CTSS]{CTSS()}} object
wrapping genomic ranges.  A \code{filteredCTSSidx} column metadata will be present
if \code{\link[=clusterCTSS]{clusterCTSS()}} was ran earlier.
}
\description{
Extracts the genomic coordinates of all detected TSSs
from \link{CAGEset} and \link{CAGEexp} objects.
}
\examples{
CTSS <- CTSScoordinates(exampleCAGEset)
head(CTSS)

CTSScoordinatesGR(exampleCAGEset)

}
\seealso{
\code{\link{getCTSS}}

\code{\link{clusterCTSS}}

Other CAGEr accessor methods: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CTSSnormalizedTpm}},
  \code{\link{CTSStagCountTable}},
  \code{\link{CTSStagCount}}, \code{\link{GeneExpDESeq2}},
  \code{\link{GeneExpSE}},
  \code{\link{consensusClustersGR}},
  \code{\link{genomeName}}, \code{\link{inputFilesType}},
  \code{\link{inputFiles}}, \code{\link{librarySizes}},
  \code{\link{sampleLabels}},
  \code{\link{seqNameTotalsSE}}, \code{\link{tagClusters}}
}
\author{
Vanja Haberle

Charles Plessy
}
\concept{CAGEr accessor methods}
