% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R, R/SetMethods.R
\name{CTSScumulativesTagClusters}
\alias{CTSScumulativesTagClusters}
\alias{`CTSScumulativesTagClusters<-`}
\title{Get/set CTSS cumulative TC data}
\usage{
CTSScumulativesTagClusters(object)

CTSScumulativesTagClusters(object) <- value
}
\arguments{
\item{object}{A \code{\link{CAGEset}} or \code{\link{CAGEset}} object.}

\item{value}{CTSScumulativesTagClusters data}
}
\seealso{
Other CAGEr clusters functions: \code{\link{aggregateTagClusters}},
  \code{\link{clusterCTSS}},
  \code{\link{consensusClustersDESeq2}},
  \code{\link{consensusClusters}},
  \code{\link{cumulativeCTSSdistribution}},
  \code{\link{plotInterquantileWidth}},
  \code{\link{quantilePositions}},
  \code{\link{tagClusters}}
}
