% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R, R/SetMethods.R
\docType{methods}
\name{CTSScumulativesTagClusters}
\alias{CTSScumulativesTagClusters}
\alias{CTSScumulativesTagClusters,CAGEset-method}
\alias{CTSScumulativesTagClusters,CAGEexp-method}
\alias{CTSScumulativesTagClusters<-}
\alias{CTSScumulativesTagClusters<-,CAGEset-method}
\alias{CTSScumulativesTagClusters<-,CAGEexp-method}
\title{Get/set CTSS cumulative TC data}
\usage{
CTSScumulativesTagClusters(object)

\S4method{CTSScumulativesTagClusters}{CAGEset}(object)

\S4method{CTSScumulativesTagClusters}{CAGEexp}(object)

CTSScumulativesTagClusters(object) <- value

\S4method{CTSScumulativesTagClusters}{CAGEset}(object) <- value

\S4method{CTSScumulativesTagClusters}{CAGEexp}(object) <- value
}
\arguments{
\item{object}{A \code{\link{CAGEset}} or \code{\link{CAGEset}} object.}

\item{value}{CTSScumulativesTagClusters data}
}
\description{
Accessor function.
}
\seealso{
Other CAGEr clusters functions: \code{\link{aggregateTagClusters}},
  \code{\link{clusterCTSS}},
  \code{\link{consensusClustersDESeq2}},
  \code{\link{consensusClustersGR}},
  \code{\link{cumulativeCTSSdistribution}},
  \code{\link{plotInterquantileWidth}},
  \code{\link{quantilePositions}},
  \code{\link{tagClusters}}

Other CAGEr accessor methods: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScoordinates}},
  \code{\link{CTSSnormalizedTpm}},
  \code{\link{CTSStagCountTable}},
  \code{\link{CTSStagCount}}, \code{\link{GeneExpDESeq2}},
  \code{\link{GeneExpSE}}, \code{\link{genomeName}},
  \code{\link{inputFilesType}}, \code{\link{inputFiles}},
  \code{\link{librarySizes}}, \code{\link{sampleLabels}},
  \code{\link{tagClusters}}
}
