% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R, R/SetMethods.R
\docType{methods}
\name{CTSScumulativesTagClusters}
\alias{CTSScumulativesTagClusters}
\alias{CTSScumulativesTagClusters,CAGEset-method}
\alias{CTSScumulativesTagClusters,CAGEexp-method}
\alias{CTSScumulativesCC}
\alias{CTSScumulativesCC,CAGEset-method}
\alias{CTSScumulativesCC,CAGEexp-method}
\alias{CTSScumulativesTagClusters<-}
\alias{CTSScumulativesTagClusters<-,CAGEset-method}
\alias{CTSScumulativesTagClusters<-,CAGEexp-method}
\title{Get/set CTSS cumulative TC or CC data}
\usage{
CTSScumulativesTagClusters(object, samples = NULL)

\S4method{CTSScumulativesTagClusters}{CAGEset}(object, samples = NULL)

\S4method{CTSScumulativesTagClusters}{CAGEexp}(object, samples = NULL)

CTSScumulativesCC(object, samples = NULL)

\S4method{CTSScumulativesCC}{CAGEset}(object, samples = NULL)

\S4method{CTSScumulativesCC}{CAGEexp}(object, samples = NULL)

CTSScumulativesTagClusters(object) <- value

\S4method{CTSScumulativesTagClusters}{CAGEset}(object) <- value

\S4method{CTSScumulativesTagClusters}{CAGEexp}(object) <- value
}
\arguments{
\item{object}{A \code{\link{CAGEset}} or \code{\link{CAGEset}} object.}

\item{samples}{One or more valid sample names.}

\item{value}{CTSScumulativesTagClusters data}
}
\value{
List of numeric Rle.
}
\description{
Accessor function.
}
\seealso{
Other CAGEr clusters functions: \code{\link{CTSSclusteringMethod}},
  \code{\link{CustomConsensusClusters}},
  \code{\link{aggregateTagClusters}},
  \code{\link{clusterCTSS}},
  \code{\link{consensusClustersDESeq2}},
  \code{\link{consensusClustersGR}},
  \code{\link{cumulativeCTSSdistribution}},
  \code{\link{plotInterquantileWidth}},
  \code{\link{quantilePositions}},
  \code{\link{tagClusters}}

Other CAGEr accessor methods: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScoordinates}},
  \code{\link{CTSSnormalizedTpm}},
  \code{\link{CTSStagCountTable}},
  \code{\link{CTSStagCount}}, \code{\link{GeneExpDESeq2}},
  \code{\link{GeneExpSE}},
  \code{\link{consensusClustersGR}},
  \code{\link{genomeName}}, \code{\link{inputFilesType}},
  \code{\link{inputFiles}}, \code{\link{librarySizes}},
  \code{\link{sampleLabels}},
  \code{\link{seqNameTotalsSE}}, \code{\link{tagClusters}}
}
\concept{CAGEr accessor methods}
\concept{CAGEr clusters functions}
