% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\docType{methods}
\name{CTSSnormalizedTpm}
\alias{CTSSnormalizedTpm}
\title{Extracting normalized CAGE signal for TSSs from CAGEr objects}
\usage{
CTSSnormalizedTpm(object)
}
\arguments{
\item{object}{A CAGEset or CAGEexp object.}
}
\value{
Returns a \code{data.frame} with normalized CAGE signal supporting
each TSS (rows) in every CAGE dataset (columns).
}
\description{
Extracts the normalized CAGE signal for all detected TSSs
in all CAGE datasets from \code{\link{CAGEset}} and
\code{\link{CAGEexp}} objects.
}
\examples{
load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
CAGEsignal <- CTSSnormalizedTpm(exampleCAGEset)
head(CAGEsignal)

}
\seealso{
\code{\link{normalizeTagCount}}

Other CAGEr accessor methods: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScoordinates}},
  \code{\link{CTSStagCountTable}},
  \code{\link{CTSStagCount}}, \code{\link{genomeName}},
  \code{\link{inputFilesType}}, \code{\link{inputFiles}},
  \code{\link{librarySizes}}, \code{\link{sampleLabels}}
}
\author{
Vanja Haberle
}
