% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\docType{methods}
\name{CTSSnormalizedTpm}
\alias{CTSSnormalizedTpm}
\alias{CTSSnormalizedTpmDf}
\alias{CTSSnormalizedTpmDF}
\title{Extracting normalized CAGE signal for TSSs from CAGEr objects}
\usage{
CTSSnormalizedTpm(object)

CTSSnormalizedTpmDf(object)

CTSSnormalizedTpmDF(object)
}
\arguments{
\item{object}{A CAGEset or CAGEexp object.}
}
\value{
\code{CTSSnormalizedTpm} returns a \code{data.frame} containing coordinates
and normalized CAGE signal supporting each TSS (rows) in every CAGE dataset (columns).

\code{CTSSnormalizedTpmDf} returns a \code{data.frame} of normalised expression values.

\code{CTSSnormalizedTpmDF} returns a \code{DataFrame} of normalised expression values.
}
\description{
Extracts the normalized CAGE signal for all detected TSSs
in all CAGE datasets from \code{\link{CAGEset}} and
\code{\link{CAGEexp}} objects.
}
\examples{
load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
head(CTSSnormalizedTpm(exampleCAGEset))

ce <- readRDS(system.file(package = "CAGEr", "extdata/CAGEexp.rds"))
normalizeTagCount(ce)
head(CTSSnormalizedTpm(ce))

load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
head(CTSSnormalizedTpmDf(exampleCAGEset))

ce <- readRDS(system.file(package = "CAGEr", "extdata/CAGEexp.rds"))
normalizeTagCount(ce)
head(CTSSnormalizedTpmDf(ce))
}
\seealso{
\code{\link{normalizeTagCount}}

Other CAGEr accessor methods: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScoordinates}},
  \code{\link{CTSStagCountTable}},
  \code{\link{CTSStagCount}}, \code{\link{genomeName}},
  \code{\link{inputFilesType}}, \code{\link{inputFiles}},
  \code{\link{librarySizes}}, \code{\link{sampleLabels}},
  \code{\link{tagClusters}}
}
\author{
Vanja Haberle
}
