% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\docType{methods}
\name{CTSSnormalizedTpm}
\alias{CTSSnormalizedTpm}
\alias{CTSSnormalizedTpm,CAGEset-method}
\alias{CTSSnormalizedTpm,CAGEexp-method}
\alias{CTSSnormalizedTpmDf}
\alias{CTSSnormalizedTpmDf,CAGEset-method}
\alias{CTSSnormalizedTpmDf,CAGEexp-method}
\alias{CTSSnormalizedTpmDF}
\alias{CTSSnormalizedTpmDF,CAGEset-method}
\alias{CTSSnormalizedTpmDF,CAGEexp-method}
\alias{CTSSnormalizedTpmGR}
\alias{CTSSnormalizedTpmGR,CAGEr-method}
\title{Extracting normalized CAGE signal for TSSs from CAGEr objects}
\usage{
CTSSnormalizedTpm(object)

\S4method{CTSSnormalizedTpm}{CAGEset}(object)

\S4method{CTSSnormalizedTpm}{CAGEexp}(object)

CTSSnormalizedTpmDf(object)

\S4method{CTSSnormalizedTpmDf}{CAGEset}(object)

\S4method{CTSSnormalizedTpmDf}{CAGEexp}(object)

CTSSnormalizedTpmDF(object)

\S4method{CTSSnormalizedTpmDF}{CAGEset}(object)

\S4method{CTSSnormalizedTpmDF}{CAGEexp}(object)

CTSSnormalizedTpmGR(object, samples)

\S4method{CTSSnormalizedTpmGR}{CAGEr}(object, samples)
}
\arguments{
\item{object}{A CAGEset or CAGEexp object.}

\item{samples}{The name of sample(s) as reported by \code{sampleLabels(object)},
or the number identifying the sample(s).}
}
\value{
\code{CTSSnormalizedTpm} returns a \code{data.frame} containing coordinates
and normalized CAGE signal supporting each TSS (rows) in every CAGE dataset (columns).

\code{CTSSnormalizedTpmDf} returns a \code{data.frame} of normalised expression values.

\code{CTSSnormalizedTpmDF} returns a \code{DataFrame} of normalised expression values.
}
\description{
Extracts the normalized CAGE signal for all detected TSSs
in all CAGE datasets from \code{\link{CAGEset}} and
\code{\link{CAGEexp}} objects.
}
\examples{
head(CTSSnormalizedTpm(exampleCAGEset))

normalizeTagCount(exampleCAGEexp)
head(CTSSnormalizedTpm(exampleCAGEexp))

head(CTSSnormalizedTpmDf(exampleCAGEset))

normalizeTagCount(exampleCAGEexp)
head(CTSSnormalizedTpmDf(exampleCAGEexp))

}
\seealso{
\code{\link{normalizeTagCount}}

Other CAGEr accessor methods: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScoordinates}},
  \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CTSStagCountTable}},
  \code{\link{CTSStagCount}}, \code{\link{GeneExpDESeq2}},
  \code{\link{GeneExpSE}},
  \code{\link{consensusClustersGR}},
  \code{\link{genomeName}}, \code{\link{inputFilesType}},
  \code{\link{inputFiles}}, \code{\link{librarySizes}},
  \code{\link{sampleLabels}},
  \code{\link{seqNameTotalsSE}}, \code{\link{tagClusters}}
}
\author{
Vanja Haberle
}
\concept{CAGEr accessor methods}
