% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\docType{methods}
\name{CTSStagCount}
\alias{CTSStagCount}
\alias{CTSStagCountDf}
\alias{CTSStagCountDF}
\alias{CTSStagCountDA}
\alias{CTSStagCountGR}
\alias{CTSStagCountSE}
\title{Extract raw CAGE TSSs expression tables from CAGEr objects}
\usage{
CTSStagCount(object)

CTSStagCountDf(object)

CTSStagCountDF(object)

CTSStagCountDA(object)

CTSStagCountGR(object, sample)

CTSStagCountSE(object)
}
\arguments{
\item{object}{A CAGEset or CAGEexp object.}

\item{sample}{(For \code{CTSStagCountGR}.) Name or number of a sample.}
}
\value{
Returns an object with number of CAGE tags supporting each TSS
(rows) in every CAGE dataset (columns).  The class of the object depends on the
function being called:

\itemize{
  \item{\code{CTSStagCount}:}
    {A \code{\link{data.frame}}, containing coordinates and expression values.}
  \item{\code{CTSStagCountDf}:}
    {A \code{data.frame}, containing only expression values.}
  \item{\code{CTSStagCountDF}:}
    {A \code{\link{DataFrame}} of \code{\link{Rle}} integers.}
  \item{\code{CTSStagCountDA}:}
    {A \code{\link{DelayedArray}} wrapping a \code{DataFrame} of \code{Rle} integers.}
  \item{\code{CTSStagCountSE}:}
    {A \code{\link{RangedSummarizedExperiment}} containing a \code{DataFrame} of
     \code{Rle} integers.}
  \item{\code{CTSStagCountGR}:}
    {A \code{\link{GenomicRanges}} object containing a \code{score} column indicating
    expression values for a given sample.}
}
}
\description{
Extracts the tag count for all detected TSSs in all CAGE datasets
             from \code{\link{CAGEset}} and \code{\link{CAGEexp}} objects.
}
\examples{
load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
head(CTSStagCount(exampleCAGEset))
head(CTSStagCountDf(exampleCAGEset))
CTSStagCountDF(exampleCAGEset)
CTSStagCountDA(exampleCAGEset)

exampleCAGEexp <- readRDS(system.file("extdata", "CAGEexp.rds", package="CAGEr"))
head(CTSStagCount(exampleCAGEexp))
head(CTSStagCountDf(exampleCAGEset))
CTSStagCountDF(exampleCAGEset)
CTSStagCountDA(exampleCAGEset)

}
\seealso{
\code{\link{getCTSS}}

Other CAGEr accessor methods: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScoordinates}},
  \code{\link{CTSSnormalizedTpm}},
  \code{\link{CTSStagCountTable}},
  \code{\link{genomeName}}, \code{\link{inputFilesType}},
  \code{\link{inputFiles}}, \code{\link{librarySizes}},
  \code{\link{sampleLabels}}, \code{\link{tagClusters}}
}
\author{
Vanja Haberle

Charles Plessy
}
