% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\docType{methods}
\name{CTSStagCount}
\alias{CTSStagCount}
\alias{CTSStagCount,CAGEset-method}
\alias{CTSStagCount,CAGEexp-method}
\alias{CTSStagCountDf}
\alias{CTSStagCountDf,CAGEset-method}
\alias{CTSStagCountDf,CAGEexp-method}
\alias{CTSStagCountDF}
\alias{CTSStagCountDF,CAGEset-method}
\alias{CTSStagCountDF,CAGEexp-method}
\alias{CTSStagCountDA}
\alias{CTSStagCountDA,CAGEr-method}
\alias{CTSStagCountGR}
\alias{CTSStagCountGR,CAGEr-method}
\alias{CTSStagCountSE}
\title{Extract raw CAGE TSSs expression tables from \code{\link{CAGEr}} objects}
\usage{
CTSStagCount(object)

\S4method{CTSStagCount}{CAGEset}(object)

\S4method{CTSStagCount}{CAGEexp}(object)

CTSStagCountDf(object)

\S4method{CTSStagCountDf}{CAGEset}(object)

\S4method{CTSStagCountDf}{CAGEexp}(object)

CTSStagCountDF(object)

\S4method{CTSStagCountDF}{CAGEset}(object)

\S4method{CTSStagCountDF}{CAGEexp}(object)

CTSStagCountDA(object)

\S4method{CTSStagCountDA}{CAGEr}(object)

CTSStagCountGR(object, samples)

\S4method{CTSStagCountGR}{CAGEr}(object, samples)

CTSStagCountSE(object)
}
\arguments{
\item{object}{A \code{CAGEr} object.}

\item{samples}{(For \code{CTSStagCountGR}.) Name(s) or number(s) identifying sample(s).}
}
\value{
Returns an object with number of CAGE tags supporting each TSS
(rows) in every CAGE dataset (columns).  The class of the object depends on the
function being called:
\itemize{
\item \code{CTSStagCount}:   A \code{\link{data.frame}}, containing coordinates and expression values.
\item \code{CTSStagCountDf}: A `data.frame``, containing only expression values.
\item \code{CTSStagCountDF}: A \code{\link{DataFrame}} of \code{\link{Rle}} integers.
\item \code{CTSStagCountDA}: A \code{\link{DelayedArray}} wrapping a \code{DataFrame} of \code{Rle} integers.
\item \code{CTSStagCountSE}: A \code{\link{RangedSummarizedExperiment}}\code{containing a}DataFrame\code{of}Rle` integers.
\item \code{CTSStagCountGR}: A \code{CTSS} object (wrapping \code{GRanges}) containing a \code{score}
column indicating expression values for a given sample.
}
}
\description{
Extracts the tag count for all detected TSSs in all CAGE datasets
from \code{\link{CAGEset}} and \code{\link{CAGEexp}} objects.
}
\examples{
head(CTSStagCount(exampleCAGEset))
head(CTSStagCountDf(exampleCAGEset))
CTSStagCountDF(exampleCAGEset)
CTSStagCountDA(exampleCAGEset)

head(CTSStagCount(exampleCAGEexp))
head(CTSStagCountDf(exampleCAGEexp))
CTSStagCountDF(exampleCAGEset)
CTSStagCountDA(exampleCAGEset)

}
\seealso{
\code{\link[=getCTSS]{getCTSS()}}

Other CAGEr accessor methods: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScoordinates}},
  \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CTSSnormalizedTpm}},
  \code{\link{CTSStagCountTable}},
  \code{\link{GeneExpDESeq2}}, \code{\link{GeneExpSE}},
  \code{\link{consensusClustersGR}},
  \code{\link{genomeName}}, \code{\link{inputFilesType}},
  \code{\link{inputFiles}}, \code{\link{librarySizes}},
  \code{\link{sampleLabels}},
  \code{\link{seqNameTotalsSE}}, \code{\link{tagClusters}}
}
\author{
Vanja Haberle

Charles Plessy
}
\concept{CAGEr CTSS methods}
\concept{CAGEr accessor methods}
