% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotations.R
\docType{methods}
\name{CTSStoGenes}
\alias{CTSStoGenes}
\title{Make a gene expression table.}
\usage{
CTSStoGenes(object)
}
\arguments{
\item{object}{A CAGEexp object that was annotated with the \code{annotateCTSS} function.}
}
\value{
The input object with the following modifications:

\itemize{
  \item A new \code{geneExpMatrix} experiment containing gene expression levels as
    a \code{SummarizedExperiemnt} object with one assay called \code{counts}, which
    is plain \code{matrix} of integers.  (This plays better than \code{Rle DataFrames}
    when interfacing with downstream packages like DESeq2, and since the number of
    genes is limited, a \code{matrix} will not cause problems of performance.)
  \item New \code{genes} column data added, indicating total number of gene symbols
    detected per library.
}
}
\description{
Add a gene expression table in the \code{GeneExpSE} experiment slot
of an annotated \code{\link{CAGEexp}} object.  \code{\link{CAGEset}} objects are
not supported.
}
\examples{
ce <- readRDS(system.file(package = "CAGEr", "extdata/CAGEexp.rds"))
annotateCTSS(ce, readRDS(system.file("extdata/Zv9_annot.rds", package = "CAGEr")))
CTSStoGenes(ce)

}
\seealso{
\code{\link{annotateCTSS}}.

Other CAGEr object modifiers: \code{\link{annotateCTSS}},
  \code{\link{clusterCTSS}}, \code{\link{getCTSS}},
  \code{\link{normalizeTagCount}},
  \code{\link{quantilePositions}}

Other CAGEr gene expression analysis functions: \code{\link{ranges2genes}}
}
\author{
Charles Plessy
}
