% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotations.R
\docType{methods}
\name{CTSStoGenes}
\alias{CTSStoGenes}
\title{Make a gene expression table.}
\usage{
CTSStoGenes(object)
}
\arguments{
\item{object}{A CAGEexp object that was annotated with the \code{annotateCTSS} function.}
}
\value{
The input object with the following modifications:

\itemize{
  \item A new \code{GeneExpSE} experiment containing gene expression levels as
    a \code{SummarizedExperiemnt} object with one assay called \code{counts}, which
    is a \code{DataFrame} of \code{Rle} integers.
  \item New \code{genes} column data added, indicating total number of gene symbols
    detected per library.
}
}
\description{
Add a gene expression table in the \code{GeneExpSE} experiment
of an annotated \code{\link{CAGEexp}} object.
}
\examples{
\dontrun{
ce <- readRDS(system.file(package = "CAGEr", "extdata/CAGEexp.rds"))
CTSStoGenes(ce)
}

}
\seealso{
\code{\link{annotateCTSS}}.

Other CAGEr object modifiers: \code{\link{annotateCTSS}},
  \code{\link{clusterCTSS}}, \code{\link{getCTSS}},
  \code{\link{normalizeTagCount}},
  \code{\link{quantilePositions}}
}
\author{
Charles Plessy
}
