% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CumulativeDistributionMethods.R
\docType{methods}
\name{cumulativeCTSSdistribution}
\alias{cumulativeCTSSdistribution}
\alias{cumulativeCTSSdistribution,CAGEr-method}
\title{Calculating cumulative sum of CAGE signal along genomic region}
\usage{
cumulativeCTSSdistribution(object, clusters, useMulticore = FALSE,
  nrCores = NULL)

\S4method{cumulativeCTSSdistribution}{CAGEr}(object, clusters,
  useMulticore = FALSE, nrCores = NULL)
}
\arguments{
\item{object}{A \code{\link{CAGEset}} object}

\item{clusters}{Which clusters should be used.  Can be either \code{clusters = "tagClusters"}
to calculate cumulative sum along tag clusters (different set of genomic coordinates
for every CAGE experiment) or \code{clusters = "consensusClusters"} to calculate
cumulative sum along consensus clusters (same set of genomic coordinates for every
CAGE experiment).}

\item{useMulticore}{Logical, should multicore be used.  \code{useMulticore = TRUE} is
supported only on Unix-like platforms.}

\item{nrCores}{Number of cores to use when \code{useMulticore = TRUE}.  Default value
\code{NULL} uses all detected cores.}
}
\value{
The slot \code{CTSScumulativesTagClusters} (when \code{clusters = "tagClusters"}) or
\code{CTSScumulativesConsensusClusters} (when \code{clusters = "consensusClusters"}) of the
provided \code{\link{CAGEset}} object will be occupied by the list containing cumulative sum
of the CAGE signal along genomic regions per CAGE experiment.
}
\description{
Calculates cumulative sum of CAGE signal along each tag cluster or
consensus cluster in every CAGE dataset within CAGEset object.
}
\examples{
CTSScumulativesTagClusters(exampleCAGEset)[[1]][1:6]
cumulativeCTSSdistribution(object = exampleCAGEset, clusters = "tagClusters")
CTSScumulativesTagClusters(exampleCAGEset)[[1]][1:6]
cumulativeCTSSdistribution(exampleCAGEset, clusters = "consensusClusters")

clusterCTSS( exampleCAGEexp, threshold = 50, thresholdIsTpm = TRUE
           , nrPassThreshold = 1, method = "distclu", maxDist = 20
           , removeSingletons = TRUE, keepSingletonsAbove = 100)
cumulativeCTSSdistribution(exampleCAGEexp, clusters = "tagClusters")
aggregateTagClusters(exampleCAGEexp, tpmThreshold = 50, excludeSignalBelowThreshold = FALSE, maxDist = 100)
cumulativeCTSSdistribution(exampleCAGEexp, clusters = "consensusClusters")
           
}
\seealso{
Other CAGEr object modifiers: \code{\link{CTSStoGenes}},
  \code{\link{CustomConsensusClusters}},
  \code{\link{aggregateTagClusters}},
  \code{\link{annotateCTSS}}, \code{\link{clusterCTSS}},
  \code{\link{getCTSS}}, \code{\link{normalizeTagCount}},
  \code{\link{quantilePositions}},
  \code{\link{summariseChrExpr}}

Other CAGEr clusters functions: \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CustomConsensusClusters}},
  \code{\link{aggregateTagClusters}},
  \code{\link{clusterCTSS}},
  \code{\link{consensusClustersDESeq2}},
  \code{\link{consensusClustersGR}},
  \code{\link{plotInterquantileWidth}},
  \code{\link{quantilePositions}},
  \code{\link{tagClusters}}
}
\author{
Vanja Haberle
}
