% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CumulativeDistributionMethods.R
\docType{methods}
\name{cumulativeCTSSdistribution}
\alias{cumulativeCTSSdistribution}
\alias{cumulativeCTSSdistribution,CAGEr-method}
\title{Cumulative sums of CAGE counts along genomic regions}
\usage{
cumulativeCTSSdistribution(object, clusters = c("tagClusters",
  "consensusClusters"), useMulticore = FALSE, nrCores = NULL)

\S4method{cumulativeCTSSdistribution}{CAGEr}(object,
  clusters = c("tagClusters", "consensusClusters"),
  useMulticore = FALSE, nrCores = NULL)
}
\arguments{
\item{object}{A \code{\link{CAGEr}} object}

\item{clusters}{\code{tagClusters} or \code{consensusClusters}.}

\item{useMulticore}{Logical, should multicore be used.
\code{useMulticore = TRUE} has no effect on non-Unix-like platforms.}

\item{nrCores}{Number of cores to use when \code{useMulticore = TRUE}
(set to \code{NULL} to use all detected cores).}
}
\value{
For \code{CAGEset} objects, the slot \code{CTSScumulativesTagClusters}
(when \code{clusters = "tagClusters"}) or \code{CTSScumulativesConsensusClusters}
(when \code{clusters = "consensusClusters"}) of the will be occupied by the
list containing cumulative sum of the CAGE signal along genomic regions per
CAGE experiment.  In \code{CAGEexp} objects, cumulative sums are stored
in the metadata slot using the \code{RleList} class.
}
\description{
Calculates the cumulative sum of raw CAGE counts along each tag
cluster or consensus cluster in every sample within a CAGEr object.
}
\examples{
CTSScumulativesTagClusters(exampleCAGEset)[[1]][1:6]
cumulativeCTSSdistribution(object = exampleCAGEset, clusters = "tagClusters")
CTSScumulativesTagClusters(exampleCAGEset)[[1]][1:6]
cumulativeCTSSdistribution(exampleCAGEset, clusters = "consensusClusters")

clusterCTSS( exampleCAGEexp, threshold = 50, thresholdIsTpm = TRUE
           , nrPassThreshold = 1, method = "distclu", maxDist = 20
           , removeSingletons = TRUE, keepSingletonsAbove = 100)
cumulativeCTSSdistribution(exampleCAGEexp, clusters = "tagClusters")
aggregateTagClusters( exampleCAGEexp, tpmThreshold = 50
                    , excludeSignalBelowThreshold = FALSE, maxDist = 100)
cumulativeCTSSdistribution(exampleCAGEexp, clusters = "consensusClusters")
           
}
\seealso{
Other CAGEr object modifiers: \code{\link{CTSStoGenes}},
  \code{\link{CustomConsensusClusters}},
  \code{\link{aggregateTagClusters}},
  \code{\link{annotateCTSS}}, \code{\link{clusterCTSS}},
  \code{\link{getCTSS}}, \code{\link{normalizeTagCount}},
  \code{\link{quantilePositions}},
  \code{\link{summariseChrExpr}}

Other CAGEr clusters functions: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CustomConsensusClusters}},
  \code{\link{aggregateTagClusters}},
  \code{\link{clusterCTSS}},
  \code{\link{consensusClustersDESeq2}},
  \code{\link{consensusClustersGR}},
  \code{\link{plotInterquantileWidth}},
  \code{\link{quantilePositions}},
  \code{\link{tagClusters}}
}
\author{
Vanja Haberle
}
\concept{CAGEr clusters functions}
\concept{CAGEr object modifiers}
