% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationMethods.R
\docType{methods}
\name{CustomConsensusClusters}
\alias{CustomConsensusClusters}
\alias{CustomConsensusClusters,CAGEexp,GRanges-method}
\title{Expression levels of consensus cluster}
\usage{
CustomConsensusClusters(object, clusters, threshold = 0,
  nrPassThreshold = 1, thresholdIsTpm = TRUE)

\S4method{CustomConsensusClusters}{CAGEexp,GRanges}(object, clusters,
  threshold = 0, nrPassThreshold = 1, thresholdIsTpm = TRUE)
}
\arguments{
\item{object}{A \code{CAGEexp} object}

\item{clusters}{Consensus clusters in \code{\link{GRanges}} format.}

\item{threshold, nrPassThreshold}{Only CTSSs with signal \code{>= threshold} in
\code{>= nrPassThreshold} experiments will be used for clustering and will
contribute towards total signal of the cluster.}

\item{thresholdIsTpm}{Logical, is threshold raw tag count value (FALSE) or
normalized signal (TRUE).}
}
\value{
stores the result as a new \code{\link{RangedSummarizedExperiment}} in the
\code{experiment} slot of the object.  The assays of the new experiment are called
\code{counts} and \code{normalized}.  An \code{outOfClusters} column is added
to the sample metadata to reflect the number of molecules that do not have
their TSS in a consensus cluster.
}
\description{
Intersects custom consensus clusters with the CTSS data in a
\code{\link{CAGEexp}} object, and stores the result as a expression matrices
(raw and normalised tag counts).
}
\details{
Consensus clusters must not overlap, so that a single base of the
genome can only be attributed to a single cluster.  This is enforced by the
\code{\link{.ConsensusClusters}} constructor.
}
\examples{

cc <- consensusClustersGR(exampleCAGEexp)
CustomConsensusClusters(exampleCAGEexp, cc)

}
\seealso{
Other CAGEr object modifiers: \code{\link{CTSStoGenes}},
  \code{\link{aggregateTagClusters}},
  \code{\link{annotateCTSS}}, \code{\link{clusterCTSS}},
  \code{\link{cumulativeCTSSdistribution}},
  \code{\link{getCTSS}}, \code{\link{normalizeTagCount}},
  \code{\link{quantilePositions}},
  \code{\link{summariseChrExpr}}

Other CAGEr clusters functions: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScumulativesTagClusters}},
  \code{\link{aggregateTagClusters}},
  \code{\link{clusterCTSS}},
  \code{\link{consensusClustersDESeq2}},
  \code{\link{consensusClustersGR}},
  \code{\link{cumulativeCTSSdistribution}},
  \code{\link{plotInterquantileWidth}},
  \code{\link{quantilePositions}},
  \code{\link{tagClusters}}
}
\author{
Charles Plessy
}
\concept{CAGEr clusters functions}
\concept{CAGEr object modifiers}
