% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distclu.R
\name{distclu}
\alias{distclu}
\alias{distclu,SummarizedExperiment-method}
\alias{distclu,CTSS-method}
\title{Distance clustering}
\usage{
distclu(object, max.dist = 20, keepSingletonsAbove = 0)

\S4method{distclu}{SummarizedExperiment}(object, max.dist = 20, keepSingletonsAbove = 0)

\S4method{distclu}{CTSS}(object, max.dist = 20, keepSingletonsAbove = 0)
}
\arguments{
\item{object}{The \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::RangedSummarizedExperiment}} object
containing CTSS information, or just a \code{\link{CTSS}} object.}

\item{max.dist}{Maximal distance between two neighbouring CTSSs for them to
be part of the same cluster.}

\item{keepSingletonsAbove}{Remove "singleton" tag clusters of width 1 with
signal \verb{< keepSingletonsAbove}.  Default value \code{0} results in keeping
all TCs by default.  Setting it to \code{Inf} removes all singletons.}
}
\value{
For \code{CTSS} input, a \code{\link{TagClusters}} object, and for
\code{SummarizedExperiment} in put, a \code{\link{GRangesList}} of \code{\link{TagClusters}} objects.
}
\description{
Distance clustering
}
\examples{
distclu(CTSSnormalizedTpmGR(exampleCAGEexp, 1)[1:10])
distclu(CTSStagCountSE(exampleCAGEexp)[1:25,])

}
\seealso{
Other CAGEr clustering methods: 
\code{\link{consensusClustersTpm}()},
\code{\link{paraclu}()}
}
\concept{CAGEr clustering methods}
