% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringFunctions.R
\name{distclu}
\alias{distclu}
\alias{distclu,SummarizedExperiment-method}
\alias{distclu,CTSS-method}
\title{Distance clustering}
\usage{
distclu(
  object,
  max.dist = 20,
  removeSingletons = FALSE,
  keepSingletonsAbove = Inf
)

\S4method{distclu}{SummarizedExperiment}(
  object,
  max.dist = 20,
  removeSingletons = FALSE,
  keepSingletonsAbove = Inf
)

\S4method{distclu}{CTSS}(
  object,
  max.dist = 20,
  removeSingletons = FALSE,
  keepSingletonsAbove = Inf
)
}
\arguments{
\item{object}{The \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::RangedSummarizedExperiment}} object
containing CTSS information.}

\item{max.dist}{Maximal distance between two neighbouring CTSSs for them to
be part of the same cluster.}

\item{removeSingletons}{Logical indicating if tag clusters containing only
one CTSS be removed.}

\item{keepSingletonsAbove}{Controls which singleton tag clusters will be
removed.  When \code{removeSingletons = TRUE}, only singletons with signal
\verb{< keepSingletonsAbove} will be removed.  Useful to prevent removing
highly supported singleton tag clusters.  Default value \code{Inf} results
in removing all singleton TCs when \code{removeSingletons = TRUE}.}
}
\description{
Distance clustering
}
\examples{
distclu(CTSSnormalizedTpmGR(exampleCAGEexp, 1)[1:10])
distclu(CTSStagCountSE(exampleCAGEexp)[1:25,])

}
\seealso{
Other CAGEr clustering methods: 
\code{\link{consensusClustersTpm}()},
\code{\link{paraclu}()}
}
\concept{CAGEr clustering methods}
