% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\name{expressionClasses}
\alias{expressionClasses}
\alias{expressionClasses,CAGEexp-method}
\title{Extract labels of expression classes}
\usage{
expressionClasses(object, what = c("CTSS", "consensusClusters"))

\S4method{expressionClasses}{CAGEexp}(object, what = c("CTSS", "consensusClusters"))

\S4method{expressionClasses}{CAGEexp}(object, what = c("CTSS", "consensusClusters"))
}
\arguments{
\item{object}{A \code{\link{CAGEexp}} object.}

\item{what}{Which level of expression clustering should be used. Can be either
\code{"CTSS"} to extract labels of expression classes of individual CTSSs or
\code{"consensusClusters"} to extract labels of expression classes of consensus
clusters.}
}
\value{
Returns character vector of labels of expression classes.  The number of labels
matches the number of expression clusters returned by \code{\link{getExpressionProfiles}}
function.
}
\description{
Retrieves labels of expression classes of either individual CTSSs or consensus
clusters from a CAGEexp object.
}
\examples{

exprClasses <- expressionClasses(exampleCAGEexp, what = "CTSS")
exprClasses

}
\seealso{
\code{\link{getExpressionProfiles}} \code{\link{plotExpressionProfiles}}
\code{\link{extractExpressionClass}}
}
