% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\name{filteredCTSSidx}
\alias{filteredCTSSidx}
\alias{filteredCTSSidx,CAGEexp-method}
\title{The \code{filteredCTSSidx()} function is in \emph{CAGEr} functions to retrieve the
result of the \code{flagLowExpCTSS()} function in a safe way.}
\usage{
filteredCTSSidx(object)

\S4method{filteredCTSSidx}{CAGEexp}(object)
}
\arguments{
\item{object}{A \code{\link{CAGEexp}} object}
}
\value{
Returns the value of \code{filteredCTSSidx} in the row ranges of the
tag count matrix experiment of the \code{CAGEexp} object, or \code{Rle(TRUE)} if it
was \code{NULL}
}
\description{
The \code{filteredCTSSidx()} function is in \emph{CAGEr} functions to retrieve the
result of the \code{flagLowExpCTSS()} function in a safe way.
}
\examples{

filteredCTSSidx(exampleCAGEexp)

}
\seealso{
Other CAGEr filter functions: 
\code{\link{flagByUpstreamSequences}()},
\code{\link{flagLowExpCTSS}()}

Other CAGEr accessor methods: 
\code{\link{.export_tag_counts}()},
\code{\link{CTSScoordinatesGR}()},
\code{\link{CTSScumulativesTagClusters}()},
\code{\link{CTSSnormalizedTpmDF}()},
\code{\link{CTSStagCountDF}()},
\code{\link{GeneExpDESeq2}()},
\code{\link{GeneExpSE}()},
\code{\link{consensusClustersGR}()},
\code{\link{exportNormalizedCTSS}()},
\code{\link{expressionClasses}()},
\code{\link{genomeName}()},
\code{\link{inputFiles}()},
\code{\link{inputFilesType}()},
\code{\link{librarySizes}()},
\code{\link{sampleLabels}()},
\code{\link{seqNameTotalsSE}()},
\code{\link{tagClustersGR}()}
}
\concept{CAGEr accessor methods}
\concept{CAGEr filter functions}
