% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\docType{methods}
\name{GeneExpSE}
\alias{GeneExpSE}
\alias{GeneExpSE,CAGEset-method}
\alias{GeneExpSE,CAGEexp-method}
\title{Retreives the SummarizedExperiment containing gene expression levels.}
\usage{
GeneExpSE(object)

\S4method{GeneExpSE}{CAGEset}(object)

\S4method{GeneExpSE}{CAGEexp}(object)
}
\arguments{
\item{object}{A \code{\link{CAGEexp}} object.}
}
\description{
Get or set a \code{SummarizedExperiment} using the gene expression
data in the experiment slot \code{geneExpMatrix} and the sample metadata
of the \code{\link{CAGEexp}} object.
}
\examples{
GeneExpSE(exampleCAGEexp)

}
\seealso{
Other CAGEr accessor methods: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScoordinates}},
  \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CTSSnormalizedTpm}},
  \code{\link{CTSStagCountTable}},
  \code{\link{CTSStagCount}}, \code{\link{GeneExpDESeq2}},
  \code{\link{consensusClustersGR}},
  \code{\link{genomeName}}, \code{\link{inputFilesType}},
  \code{\link{inputFiles}}, \code{\link{librarySizes}},
  \code{\link{sampleLabels}},
  \code{\link{seqNameTotalsSE}}, \code{\link{tagClusters}}
}
\author{
Charles Plessy
}
\concept{CAGEr accessor methods}
