% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\docType{methods}
\name{genomeName}
\alias{genomeName}
\title{Extracting genome name from CAGEset and CAGEexp objects

Extracts the name of a referent genome from a \code{\link{CAGEset}}
and \code{\link{CAGEexp}} objects.}
\usage{
genomeName(object)
}
\arguments{
\item{object}{A CAGEset or CAGEexp object.}
}
\value{
Returns a name of a BSgenome package used as a referent genome.
}
\examples{

load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
genomeName(exampleCAGEset)

}
\author{
Vanja Haberle
}
