% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R, R/SetMethods.R
\docType{methods}
\name{genomeName}
\alias{genomeName}
\alias{genomeName<-}
\alias{genomeName<-,CAGEset-method}
\alias{genomeName<-,CAGEexp-method}
\title{Extracting genome name from CAGEr objects}
\usage{
genomeName(object)

genomeName(object) <- value

\S4method{genomeName}{CAGEset}(object) <- value

\S4method{genomeName}{CAGEexp}(object) <- value
}
\arguments{
\item{object}{A CAGEset or CAGEexp object.}

\item{value}{The name of a \code{BSgenome} package.}
}
\value{
Returns a name of a BSgenome package used as a referent genome.
}
\description{
Extracts the name of a referent genome from a
\code{\link{CAGEset}} and \code{\link{CAGEexp}} objects.
}
\details{
\code{\link{CAGEexp}} objects constructed with \code{NULL} in place
of the genome name can not run some commands that need access to genomic data,
such as BigWig export or G-correction.
}
\examples{
load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
genomeName(exampleCAGEset)

}
\seealso{
Other CAGEr accessor methods: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScoordinates}},
  \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CTSSnormalizedTpm}},
  \code{\link{CTSStagCountTable}},
  \code{\link{CTSStagCount}}, \code{\link{inputFilesType}},
  \code{\link{inputFiles}}, \code{\link{librarySizes}},
  \code{\link{sampleLabels}}, \code{\link{tagClusters}}

Other CAGEr setter methods: \code{\link{inputFilesType}},
  \code{\link{inputFiles}}, \code{\link{sampleLabels}},
  \code{\link{setColors}}
}
\author{
Vanja Haberle

Charles Plessy
}
