% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\docType{methods}
\name{genomeName}
\alias{genomeName}
\title{Extracting genome name from CAGEr objects}
\usage{
genomeName(object)
}
\arguments{
\item{object}{A CAGEset or CAGEexp object.}
}
\value{
Returns a name of a BSgenome package used as a referent genome.
}
\description{
Extracts the name of a referent genome from a
\code{\link{CAGEset}} and \code{\link{CAGEexp}} objects.
}
\examples{
load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
genomeName(exampleCAGEset)

}
\seealso{
Other CAGEr accessor methods: \code{\link{inputFilesType}},
  \code{\link{inputFiles}}, \code{\link{librarySizes}},
  \code{\link{sampleLabels}}
}
\author{
Vanja Haberle
}
