% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportMethods.R
\name{import.bam.ctss}
\alias{import.bam.ctss}
\title{import.bam.ctss}
\usage{
import.bam.ctss(filepath, filetype, sequencingQualityThreshold,
  mappingQualityThreshold, removeFirstG, correctSystematicG, genome)
}
\arguments{
\item{filepath}{The path to the BAM file.}

\item{filetype}{bam or bamPairedEnd.}

\item{sequencingQualityThreshold}{See getCTSS().}

\item{mappingQualityThreshold}{See getCTSS().}

\item{removeFirstG}{See getCTSS().}

\item{correctSystematicG}{See getCTSS().}

\item{genome}{See coerceInBSgenome().}
}
\value{
Returns a \link{CTSS} object.
}
\description{
Imports CTSS data from a BAM file.
}
\seealso{
Other loadFileIntoGPos: \code{\link{bam2CTSS}},
  \code{\link{import.CTSS}}, \code{\link{import.bam}},
  \code{\link{import.bedCTSS}},
  \code{\link{import.bedScore}},
  \code{\link{import.bedmolecule}},
  \code{\link{loadFileIntoGPos}},
  \code{\link{moleculesGR2CTSS}}
}
\concept{loadFileIntoGPos}
