% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportMethods.R
\name{import.bam}
\alias{import.bam}
\title{import.bam}
\usage{
import.bam(filepath, filetype, sequencingQualityThreshold = 10,
  mappingQualityThreshold = 20)
}
\arguments{
\item{filepath}{The path to the BAM file.}

\item{filetype}{bam or bamPairedEnd.}

\item{sequencingQualityThreshold}{See getCTSS().}

\item{mappingQualityThreshold}{See getCTSS().}
}
\description{
Imports CTSS data from a BAM file.
}
\examples{
# TODO: add exmaple file
# import.bam(system.file("extdata", "example.bam", package = "CAGEr"))
}
\seealso{
Other loadFileIntoGPos: \code{\link{bam2CTSS}},
  \code{\link{import.CTSS}}, \code{\link{import.bam.ctss}},
  \code{\link{import.bedCTSS}},
  \code{\link{import.bedScore}},
  \code{\link{import.bedmolecule}},
  \code{\link{loadFileIntoGPos}},
  \code{\link{moleculesGR2CTSS}}
}
\concept{loadFileIntoGPos}
