% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportMethods.R
\name{import.bedmolecule}
\alias{import.bedmolecule}
\title{import.bedmolecule}
\usage{
import.bedmolecule(filepath)
}
\arguments{
\item{filepath}{The path to the BED file.}
}
\value{
Returns a \code{\link{CTSS}} object.
}
\description{
Imports a BED file where each line counts for one molecule in a
GRanges object where each line represents one nucleotide.
}
\examples{
# TODO: add exmaple file
# import.BED(system.file("extdata", "example.bed", package = "CAGEr"))
}
\seealso{
Other loadFileIntoGPos: \code{\link{bam2CTSS}},
  \code{\link{import.CTSS}}, \code{\link{import.bam.ctss}},
  \code{\link{import.bam}}, \code{\link{import.bedCTSS}},
  \code{\link{import.bedScore}},
  \code{\link{loadFileIntoGPos}},
  \code{\link{moleculesGR2CTSS}}
}
\concept{loadFileIntoGPos}
