% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportMethods.R
\name{import.bedScore}
\alias{import.bedScore}
\title{import.bedScore}
\usage{
import.bedScore(filepath)
}
\arguments{
\item{filepath}{The path to the BED file.}
}
\value{
A GRanges object where each line represents one nucleotide.
}
\description{
Imports a BED file where the score indicates a number of counts
for a given alignment.
}
\examples{
# TODO: add exmaple file
# import.bedScore(system.file("extdata", "example.bed", package = "CAGEr"))
}
\seealso{
loadFileIntoGRanges

Other loadFileIntoGRanges: \code{\link{import.CTSS}},
  \code{\link{import.bam}}, \code{\link{import.bedCTSS}},
  \code{\link{import.bedmolecule}},
  \code{\link{loadFileIntoGRanges}}
}
