% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R, R/SetMethods.R
\docType{methods}
\name{inputFiles}
\alias{inputFiles}
\alias{inputFiles,CAGEset-method}
\alias{inputFiles,CAGEexp-method}
\alias{inputFiles<-}
\alias{inputFiles<-,CAGEset-method}
\alias{inputFiles<-,CAGEexp-method}
\title{Extracting paths to input files from CAGEr objects}
\usage{
inputFiles(object)

\S4method{inputFiles}{CAGEset}(object)

\S4method{inputFiles}{CAGEexp}(object)

inputFiles(object) <- value

\S4method{inputFiles}{CAGEset}(object) <- value

\S4method{inputFiles}{CAGEexp}(object) <- value
}
\arguments{
\item{object}{A CAGEset or CAGEexp object.}

\item{value}{A character vector with one file path per sample.}
}
\value{
Returns a character vector of paths to CAGE data input files.
}
\description{
Extracts the paths to CAGE data input files from
\code{\link{CAGEset}} and \code{\link{CAGEexp}} objects.
}
\examples{
inputFiles(exampleCAGEset)

}
\seealso{
Other CAGEr accessor methods: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScoordinates}},
  \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CTSSnormalizedTpm}},
  \code{\link{CTSStagCountTable}},
  \code{\link{CTSStagCount}}, \code{\link{GeneExpDESeq2}},
  \code{\link{GeneExpSE}},
  \code{\link{consensusClustersGR}},
  \code{\link{genomeName}}, \code{\link{inputFilesType}},
  \code{\link{librarySizes}}, \code{\link{sampleLabels}},
  \code{\link{seqNameTotalsSE}}, \code{\link{tagClusters}}

Other CAGEr setter methods: \code{\link{genomeName}},
  \code{\link{inputFilesType}}, \code{\link{sampleLabels}},
  \code{\link{setColors}}
}
\author{
Vanja Haberle

Charles Plessy
}
\concept{CAGEr accessor methods}
\concept{CAGEr setter methods}
