% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\docType{methods}
\name{inputFilesType}
\alias{inputFilesType}
\title{Extracting type of input files from CAGEr objects}
\usage{
inputFilesType(object)
}
\arguments{
\item{object}{A CAGEset or CAGEexp object.}
}
\value{
Returns the label of the file type of CAGE data input files,
\emph{e.g.} \code{"bam"} or \code{"ctss"}.  In the case of \code{CAGEexp}
objects, the return value is character vector with one member per sample.
}
\description{
Extracts the information on the type of CAGE data input
files from \code{\link{CAGEset}} and \code{\link{CAGEexp}} objects.
}
\examples{
load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
inputFilesType(exampleCAGEset)

}
\seealso{
Other CAGEr accessor methods: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScoordinates}},
  \code{\link{CTSSnormalizedTpm}},
  \code{\link{CTSStagCountTable}},
  \code{\link{CTSStagCount}}, \code{\link{genomeName}},
  \code{\link{inputFiles}}, \code{\link{librarySizes}},
  \code{\link{sampleLabels}}
}
\author{
Vanja Haberle
}
