% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\docType{methods}
\name{librarySizes}
\alias{librarySizes}
\title{Extracting library sizes from CAGEr objects}
\usage{
librarySizes(object)
}
\arguments{
\item{object}{A CAGEset or CAGEexp object.}
}
\value{
Returns an integer vector of total number of CAGE tags (library size) for all CAGE
datasets in the CAGEr object.
}
\description{
Extracts the library sizes (total number of CAGE tags) for all CAGE datasets
from \code{\link{CAGEset}} and \code{\link{CAGEexp}} objects.
}
\examples{
load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
librarySizes(exampleCAGEset)

}
\seealso{
Other CAGEr accessor methods: \code{\link{genomeName}},
  \code{\link{inputFilesType}}, \code{\link{inputFiles}},
  \code{\link{sampleLabels}}
}
\author{
Vanja Haberle
}
