% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationMethods.R
\docType{methods}
\name{plotCorrelation}
\alias{plotCorrelation}
\alias{plotCorrelation,CAGEr-method}
\alias{plotCorrelation2}
\alias{plotCorrelation2,CAGEr-method}
\title{Pairwise scatter plots and correlations of CAGE signal}
\usage{
plotCorrelation(object, what = c("CTSS", "consensusClusters"),
  values = c("raw", "normalized"), samples = "all", method = "pearson",
  tagCountThreshold = 1, applyThresholdBoth = FALSE, plotSize = 800)

\S4method{plotCorrelation}{CAGEr}(object, what = c("CTSS",
  "consensusClusters"), values = c("raw", "normalized"), samples = "all",
  method = "pearson", tagCountThreshold = 1, applyThresholdBoth = FALSE,
  plotSize = 800)

plotCorrelation2(object, what = c("CTSS", "consensusClusters"),
  values = c("raw", "normalized"), samples = "all", method = "pearson",
  tagCountThreshold = 1, applyThresholdBoth = FALSE)

\S4method{plotCorrelation2}{CAGEr}(object, what = c("CTSS",
  "consensusClusters"), values = c("raw", "normalized"), samples = "all",
  method = "pearson", tagCountThreshold = 1, applyThresholdBoth = FALSE)
}
\arguments{
\item{object}{A \code{\link{CAGEr}} object.}

\item{what}{The clustering level to be used for plotting and calculating
correlations.  Can be either \code{"CTSS"} to use individual TSSs or
\code{"consensusClusters"} to use consensus clusters, \emph{i.e.} entire
promoters.}

\item{values}{Use either \code{"raw"} (default) or \code{"normalized"} CAGE
signal.}

\item{samples}{Character vector indicating which samples to use.  Can be
either \code{"all"} to select all samples in a \code{CAGEr} object, or a
subset of valid sample labels as returned by the
\code{\link{sampleLabels}} function.}

\item{method}{A character string indicating which correlation coefficient
should be computed.  Passed to \code{cor} function.  Can be one of
\code{"pearson"}, \code{"spearman"}, or \code{"kendall"}.}

\item{tagCountThreshold}{Only TSSs with tag count \code{>= tagCountThreshold}
in either one (\code{applyThresholdBoth = FALSE}) or both samples
(\code{applyThresholdBoth = TRUE}) are plotted and used to calculate
correlation.}

\item{applyThresholdBoth}{See \code{tagCountThreshold} above.}

\item{plotSize}{Size of the individual comparison plot in pixels - the
total size of the resulting png will be \code{length(samples) * plotSize}
in both dimensions.  Ignored in \code{plotCorrelation2}.}
}
\value{
Displays the plot and returns a \code{matrix} of pairwise
correlations between selected samples.  The scatterplots of
\code{plotCorrelation} are colored according to the density of points, and
in \code{plotCorrelation2} they are just black and white, which is much
faster to plot.  Note that while the scatterplots are on a logarithmic scale
with pseudocount added to the zero values, the correlation coefficients are
calculated on untransformed (but thresholded) data.
}
\description{
Calculates the pairwise correlation between samples and creates
a plot matrix showing the correlation coeficients in the upper triangle, the
sample names in the diagonal, and the catter plots in the lower triangle.
}
\details{
Internally, the data is converted to a \code{data.frame} format,
so there may be performance issues if there are many samples.  On the other
hand, this kind of plot does not make much sense for large numbers of samples.

In the scatter plots, a pseodo-count equal to half the lowest score is added
to the null values so that they can appear despite logarithmic scale.
}
\examples{
plotCorrelation(object = exampleCAGEset)

plotCorrelation2(exampleCAGEexp, what = "consensusClusters", value = "normalized")

}
\seealso{
Other CAGEr plot functions: \code{\link{hanabiPlot}},
  \code{\link{plotAnnot}},
  \code{\link{plotExpressionProfiles}},
  \code{\link{plotInterquantileWidth}},
  \code{\link{plotReverseCumulatives}}
}
\author{
Vanja Haberle

Charles Plessy
}
