% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelationMethods.R
\docType{methods}
\name{plotCorrelation}
\alias{plotCorrelation}
\alias{plotCorrelation,CAGEr-method}
\alias{plotCorrelation2}
\alias{plotCorrelation2,CAGEset-method}
\alias{plotCorrelation2,CAGEexp-method}
\alias{plotCorrelation2,SummarizedExperiment-method}
\alias{plotCorrelation2,DataFrame-method}
\alias{plotCorrelation2,data.frame-method}
\alias{plotCorrelation2,matrix-method}
\title{Pairwise scatter plots and correlations of CAGE signal}
\usage{
plotCorrelation(object, what = c("CTSS", "consensusClusters"),
  values = c("raw", "normalized"), samples = "all",
  method = "pearson", tagCountThreshold = 1,
  applyThresholdBoth = FALSE, plotSize = 800)

\S4method{plotCorrelation}{CAGEr}(object, what = c("CTSS",
  "consensusClusters"), values = c("raw", "normalized"),
  samples = "all", method = "pearson", tagCountThreshold = 1,
  applyThresholdBoth = FALSE, plotSize = 800)

plotCorrelation2(object, what = c("CTSS", "consensusClusters"),
  values = c("raw", "normalized"), samples = "all",
  method = "pearson", tagCountThreshold = 1,
  applyThresholdBoth = FALSE, digits = 3)

\S4method{plotCorrelation2}{CAGEset}(object, what = c("CTSS",
  "consensusClusters"), values = c("raw", "normalized"),
  samples = "all", method = "pearson", tagCountThreshold = 1,
  applyThresholdBoth = FALSE, digits = 3)

\S4method{plotCorrelation2}{CAGEexp}(object, what = c("CTSS",
  "consensusClusters"), values = c("raw", "normalized"),
  samples = "all", method = "pearson", tagCountThreshold = 1,
  applyThresholdBoth = FALSE, digits = 3)

\S4method{plotCorrelation2}{SummarizedExperiment}(object,
  what = c("CTSS", "consensusClusters"), values = c("raw",
  "normalized"), samples = "all", method = "pearson",
  tagCountThreshold = 1, applyThresholdBoth = FALSE, digits = 3)

\S4method{plotCorrelation2}{DataFrame}(object, what = c("CTSS",
  "consensusClusters"), values = c("raw", "normalized"),
  samples = "all", method = "pearson", tagCountThreshold = 1,
  applyThresholdBoth = FALSE, digits = 3)

\S4method{plotCorrelation2}{data.frame}(object, what = c("CTSS",
  "consensusClusters"), values = c("raw", "normalized"),
  samples = "all", method = "pearson", tagCountThreshold = 1,
  applyThresholdBoth = FALSE, digits = 3)

\S4method{plotCorrelation2}{matrix}(object, what = c("CTSS",
  "consensusClusters"), values = c("raw", "normalized"),
  samples = "all", method = "pearson", tagCountThreshold = 1,
  applyThresholdBoth = FALSE, digits = 3)
}
\arguments{
\item{object}{A \code{\link{CAGEr}} object or (only for
\code{plotCorrelation2}) a \code{\link{SummarizedExperiment}} or an
expression table as a \code{\link{DataFrame}}, \code{\link{data.frame}} or
\code{\link{matrix}} object.}

\item{what}{The clustering level to be used for plotting and calculating
correlations.  Can be either \code{"CTSS"} to use individual TSSs or
\code{"consensusClusters"} to use consensus clusters, \emph{i.e.} entire
promoters.  Ignored for anything else than \code{CAGEr} objects.}

\item{values}{Use either \code{"raw"} (default) or \code{"normalized"} CAGE
signal.  Ignored for plain expression tables.}

\item{samples}{Character vector indicating which samples to use.  Can be
either \code{"all"} to select all samples in a \code{CAGEr} object, or a
subset of valid sample labels as returned by the
\code{\link{sampleLabels}} function.}

\item{method}{A character string indicating which correlation coefficient
should be computed.  Passed to \code{cor} function.  Can be one of
\code{"pearson"}, \code{"spearman"}, or \code{"kendall"}.}

\item{tagCountThreshold}{Only TSSs with tag count \code{>= tagCountThreshold}
in either one (\code{applyThresholdBoth = FALSE}) or both samples
(\code{applyThresholdBoth = TRUE}) are plotted and used to calculate
correlation.}

\item{applyThresholdBoth}{See \code{tagCountThreshold} above.}

\item{plotSize}{Size of the individual comparison plot in pixels - the
total size of the resulting png will be \code{length(samples) * plotSize}
in both dimensions.  Ignored in \code{plotCorrelation2}.}

\item{digits}{The number of significant digits for the data to be kept in log
scale.  Ignored in \code{plotCorrelation}.  In \code{plotCorrelation2}, the
number of points plotted is considerably reduced by rounding the point
coordinates to a small number of significant digits before removing
duplicates.  Chose a value that makes the plot visually indistinguishable
with non-deduplicated data, by making tests on a subset of the data.}
}
\value{
Displays the plot and returns a \code{matrix} of pairwise
correlations between selected samples.  The scatterplots of
\code{plotCorrelation} are colored according to the density of points, and
in \code{plotCorrelation2} they are just black and white, which is much
faster to plot.  Note that while the scatterplots are on a logarithmic scale
with pseudocount added to the zero values, the correlation coefficients are
calculated on untransformed (but thresholded) data.
}
\description{
Calculates the pairwise correlation between samples and creates
a plot matrix showing the correlation coeficients in the upper triangle, the
sample names in the diagonal, and the catter plots in the lower triangle.
}
\details{
In the scatter plots, a pseudo-count equal to half the lowest score
is added to the null values so that they can appear despite logarithmic scale.

\code{SummarizedExperiment} objects are expected to contain raw tag counts
in a \dQuote{counts} assay and the normalized expression scores in a
\dQuote{normalized} assay.

Avoid using large \code{matrix} objects as they are coerced to
\code{DataFrame} class without special care for efficiency.

\code{plotCorrelation2} speeds up the plotting by a) deduplicating
that data: no point is plot twice at the same coordinates, b) rounding the
data so that indistinguishable positions are plotted only once, c) using a
black square glyph for the points, d) caching some calculations that are
made repeatedly (to determine where to plot the correlation coefficients),
and e) preventing coercion of \code{DataFrames} to \code{data.frames}.
}
\examples{

plotCorrelation2(exampleCAGEexp, what = "consensusClusters", value = "normalized")

}
\seealso{
Other CAGEr plot functions: \code{\link{hanabiPlot}},
  \code{\link{plotAnnot}},
  \code{\link{plotExpressionProfiles}},
  \code{\link{plotInterquantileWidth}},
  \code{\link{plotReverseCumulatives}}
}
\author{
Vanja Haberle

Charles Plessy
}
\concept{CAGEr plot functions}
