% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportMethods.R
\docType{methods}
\name{plotExpressionProfiles}
\alias{plotExpressionProfiles}
\alias{plotExpressionProfiles,CAGEset-method}
\title{Plotting expression profiles derived from CAGE data}
\usage{
plotExpressionProfiles(object, what)

\S4method{plotExpressionProfiles}{CAGEset}(object, what)
}
\arguments{
\item{object}{A \code{\link{CAGEset}} object.}

\item{what}{Which level of expression clustering should be used.  Can be either
\code{"CTSS"} to plot expression profiles of individual CTSSs or \code{"consensusClusters"}
to plot expression profiles of consensus clusters.}
}
\description{
Creates a chart with beanplots representing distribution of normalized
expression across CAGE experiments for individual expression classes.  Different
expression classes are shown in different colors and are labeled according to the labels
returned by expression clustering.
}
\details{
The created file contains beanplots representing distribution of normalized
expression across CAGE experiments for individual expression classes shown in separate
boxes.  Each labeled box represents one expression class and contains a set of
beanplots - one per CAGE experiment. Individual CAGE experiments are shown on X-axis and
scaled normalized expression on Y-axis.  Individual beanplots show distribution of
normalized expression values of elements belonging to specific expression class in
particular CAGE experiment, and the entire box represents single expression profile.
Different expression classes (boxes) are plotted in different colors and are labeled with
labels returned by expression clustering.
}
\examples{
plotExpressionProfiles(object = exampleCAGEset, what = "CTSS")

}
\seealso{
\code{\link{getExpressionProfiles}}, \code{\link{expressionClasses}},
\code{\link{extractExpressionClass}}.

Other CAGEr plot functions: \code{\link{hanabiPlot}},
  \code{\link{plotAnnot}}, \code{\link{plotCorrelation}},
  \code{\link{plotInterquantileWidth}},
  \code{\link{plotReverseCumulatives}}
}
\author{
Vanja Haberle
}
\concept{CAGEr plot functions}
