% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileWidthMethods.R
\docType{methods}
\name{quantilePositions}
\alias{quantilePositions}
\alias{quantilePositions,CAGEr-method}
\title{Determine CTSS quantile positions within clusters}
\usage{
quantilePositions(object, clusters, qLow = 0.1, qUp = 0.9,
  useMulticore = FALSE, nrCores = NULL)

\S4method{quantilePositions}{CAGEr}(object, clusters, qLow = 0.1, qUp = 0.9,
  useMulticore = FALSE, nrCores = NULL)
}
\arguments{
\item{object}{A \code{\link{CAGEr}} object.}

\item{clusters}{Either \code{tagClusters} or \code{consensusClusters}.}

\item{qLow, qUp}{Which \dQuote{lower} or \dQuote{upper} quantiles should be
calculated. Numeric vector of values in range \code{[0,1]}.}

\item{useMulticore}{Logical, should multicore be used.
\code{useMulticore = TRUE} has only effect on Unix-like platforms.}

\item{nrCores}{Number of cores to use when \code{useMulticore = TRUE}.
Default value \code{NULL} uses all detected cores.}
}
\value{
When \code{clusters = "tagClusters"}, the slots \code{tagClustersQuantileLow}
and \code{tagClustersQuantileUp} of a provided \code{\link{CAGEset}} object will
be occupied with the positions of specified quantiles in all tag clusters for all
CAGE datasets. When \code{clusters = "consensusClusters"} the slots
\code{consensusClustersQuantileLow} and \code{consensusClustersQuantileUp} will be
occupied by the corresponding information for consensus clusters.

In \code{\link{CAGEexp}} objects, the positions of the quantiles are defined
reliatively to the start point of their cluster, for more efficient
\code{Rle} compression.  The quantile data for \emph{tag clusters} are stored
in the \code{TagClusters} objects directly.  The quantile data for
\emph{consensus clusters} are stored in \code{\link{integer}} matrices named
\dQuote{q_\emph{x}}, where \emph{x} represents the quantile (for instance,
\code{q_0.1}), and these matrices are \emph{assays} of the
\code{consensusClusters} \code{\link{RangedSummarizedExperiment}}.
}
\description{
Calculates the positions of \dQuote{upper} and \dQuote{lower}
quantiles of CAGE signal along \emph{tag clusters} or
\emph{consensus clusters} in each sample of a  CAGEr object.
}
\details{
The position of a \dQuote{lower} quantile \code{qLow} is defined as
a point that divides the cluster into two parts, so that the 5' part contains
\code{< qLow * 100\%} of the CAGE signal of that region (\code{>= qUp * 100\%}
for \dQuote{upper} quantiles).  Therefore, when \code{qLow <= qUp}, a lower
and and uppper quantile define a central region that contains
\code{>= (qUp - qLow) * 100\%} of the CAGE signal of that cluster.  The
width of that central part is refered to as "interquantile width", which is
a more robust measure of the promoter width than the total length of the
cluster.
}
\examples{
head(cbind(
  CAGEr:::tagClustersQuantileLow(exampleCAGEset, 1),
  CAGEr:::tagClustersQuantileUp (exampleCAGEset, 1)
))
quantilePositions( object = exampleCAGEset, clusters = "tagClusters"
                 , qLow = c(0.1, 0.2), qUp = c(0.8, 0.9))
head(cbind(
  CAGEr:::tagClustersQuantileLow(exampleCAGEset, 1),
  CAGEr:::tagClustersQuantileUp (exampleCAGEset,1 )
))

cumulativeCTSSdistribution(exampleCAGEset, "consensusClusters") # Defaults in object do not fit
quantilePositions( object = exampleCAGEset, clusters = "consensusClusters"
                 , qLow = c(0.1, 0.2), qUp = c(0.8, 0.9))
                 
head(cbind(
  CAGEr:::consensusClustersQuantileLow(exampleCAGEset, 1),
  CAGEr:::consensusClustersQuantileUp (exampleCAGEset , 1)
))

quantilePositions(exampleCAGEexp, "tagClusters",       qLow = c(0.1, 0.2), qUp = c(0.8, 0.9))
tagClustersGR(exampleCAGEexp)
quantilePositions(exampleCAGEexp, "consensusClusters", qLow = c(0.1, 0.2), qUp = c(0.8, 0.9))

}
\seealso{
Other CAGEr object modifiers: \code{\link{CTSStoGenes}},
  \code{\link{CustomConsensusClusters}},
  \code{\link{aggregateTagClusters}},
  \code{\link{annotateCTSS}}, \code{\link{clusterCTSS}},
  \code{\link{cumulativeCTSSdistribution}},
  \code{\link{getCTSS}}, \code{\link{normalizeTagCount}},
  \code{\link{summariseChrExpr}}

Other CAGEr clusters functions: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CustomConsensusClusters}},
  \code{\link{aggregateTagClusters}},
  \code{\link{clusterCTSS}},
  \code{\link{consensusClustersDESeq2}},
  \code{\link{consensusClustersGR}},
  \code{\link{cumulativeCTSSdistribution}},
  \code{\link{plotInterquantileWidth}},
  \code{\link{tagClusters}}
}
\author{
Vanja Haberle
}
