% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileWidthMethods.R
\docType{methods}
\name{quantilePositions}
\alias{quantilePositions}
\alias{quantilePositions,CAGEr-method}
\title{Determining positions of CAGE signal quantiles within genomic region}
\usage{
quantilePositions(object, clusters, qLow = 0.1, qUp = 0.9,
  useMulticore = FALSE, nrCores = NULL)

\S4method{quantilePositions}{CAGEr}(object, clusters, qLow = 0.1, qUp = 0.9,
  useMulticore = FALSE, nrCores = NULL)
}
\arguments{
\item{object}{A \code{\link{CAGEset}} object}

\item{clusters}{Which clusters should be used.  Can be either \code{tagClusters}
to calculate positions of quantiles in tag clusters (different set of genomic
coordinates for every CAGE experiment) or \code{consensusClusters} to calculate
positions of quantiles in consensus clusters (same set of genomic coordinates for
every CAGE experiment).}

\item{qLow}{Which "lower" quantiles should be calculated.  It has to be a numeric
vector of values in range \code{[0,1]}.  See Details.}

\item{qUp}{Which "upper" quantiles should be calculated.  It has to be a numeric
vector of values in range \code{[0,1]}.  See Details.}

\item{useMulticore}{Logical, should multicore be used.  \code{useMulticore = TRUE}
is supported only on Unix-like platforms.}

\item{nrCores}{Number of cores to use when \code{useMulticore = TRUE}.  Default
value \code{NULL} uses all detected cores.}
}
\value{
When \code{clusters = "tagClusters"}, the slots \code{tagClustersQuantileLow}
and \code{tagClustersQuantileUp} of the provided \code{\link{CAGEset}} object will
be occupied with the positions of specified quantiles in all tag clusters for all
CAGE datasets. When \code{clusters = "consensusClusters"} the slots
\code{consensusClustersQuantileLow} and \code{consensusClustersQuantileUp} will be
occupied by the corresponding information for consensus clusters.
}
\description{
Calculates positions of quantiles of CAGE signal along tag clusters or consensus
clusters in each CAGE dataset within CAGEset object.  The function calculates
positions of both \dQuote{lower} and \dQuote{upper} quantiles as described in Details.
}
\details{
Position of the "lower" quantile \code{qLow} is defined as a point that divides the
genomic region into two parts, so that the 5' part contains \code{< qLow * 100\%} of
the CAGE signal of that region. Accordingly, position of the "upper" quantile
\code{qUp} is defined as a point that divides the genomic region into two parts so
that the 5' part contains \code{>= qUp * 100\%} of the CAGE signal of that region.
Positions of one "lower" and one "upper" quantile (when \code{qLow <= qUp}) define a
central part of the genomic region that contains \code{>= (qUp - qLow) * 100\%} of
the CAGE signal of that region. Width of that central part is refered to as
"interquantile width", which is a more robust measure of the promoter width than the
total span of the region.
}
\examples{
head(cbind(
  CAGEr:::tagClustersQuantileLow(exampleCAGEset, 1),
  CAGEr:::tagClustersQuantileUp (exampleCAGEset, 1)
))
quantilePositions( object = exampleCAGEset, clusters = "tagClusters"
                 , qLow = c(0.1, 0.2), qUp = c(0.8, 0.9))
head(cbind(
  CAGEr:::tagClustersQuantileLow(exampleCAGEset, 1),
  CAGEr:::tagClustersQuantileUp (exampleCAGEset,1 )
))

cumulativeCTSSdistribution(exampleCAGEset, "consensusClusters") # Defaults in object do not fit
quantilePositions( object = exampleCAGEset, clusters = "consensusClusters"
                 , qLow = c(0.1, 0.2), qUp = c(0.8, 0.9))
                 
head(cbind(
  CAGEr:::consensusClustersQuantileLow(exampleCAGEset, 1),
  CAGEr:::consensusClustersQuantileUp (exampleCAGEset , 1)
))

quantilePositions(exampleCAGEexp, "tagClusters",       qLow = c(0.1, 0.2), qUp = c(0.8, 0.9))
tagClustersGR(exampleCAGEexp)
quantilePositions(exampleCAGEexp, "consensusClusters", qLow = c(0.1, 0.2), qUp = c(0.8, 0.9))

}
\seealso{
Other CAGEr object modifiers: \code{\link{CTSStoGenes}},
  \code{\link{CustomConsensusClusters}},
  \code{\link{aggregateTagClusters}},
  \code{\link{annotateCTSS}}, \code{\link{clusterCTSS}},
  \code{\link{cumulativeCTSSdistribution}},
  \code{\link{getCTSS}}, \code{\link{normalizeTagCount}},
  \code{\link{summariseChrExpr}}

Other CAGEr clusters functions: \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CustomConsensusClusters}},
  \code{\link{aggregateTagClusters}},
  \code{\link{clusterCTSS}},
  \code{\link{consensusClustersDESeq2}},
  \code{\link{consensusClustersGR}},
  \code{\link{cumulativeCTSSdistribution}},
  \code{\link{plotInterquantileWidth}},
  \code{\link{tagClusters}}
}
\author{
Vanja Haberle
}
