% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileWidthFunctions.R
\name{.get.quant.pos}
\alias{.get.quant.pos}
\title{Get quantile positions}
\usage{
.get.quant.pos(cumsums, clusters, q)
}
\arguments{
\item{cumsums}{Named list of vectors containing cumulative sum for each cluster
(returned by 'get.cumsum' function).}

\item{clusters}{GRanges object representing tag clusters or consensus clusters.}

\item{q}{desired quantiles - single value or a vector of values.}
}
\value{
Returns the \code{clusters} object with one more metadata column per value
in \code{q}, containing Rle integers giving the relative distance of the quantile
boundaries to the start position.
}
\description{
Function that calculates position of quantiles for CTSS clusters based on
distribution of tpm within the cluster
}
\examples{
\dontrun{ #because it runs through quantilePositions() anyway.
cumsums  <- CTSScumulativesTagClusters(object, 1)
clusters <- tagClustersGR(object, 1)
.get.quant.pos(cumsums, clusters, c(.1, .9))
}
}
