% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotations.R
\name{ranges2annot}
\alias{ranges2annot}
\title{Hierarchical annotation of CTSSes}
\usage{
ranges2annot(ranges, annot)
}
\arguments{
\item{ranges}{A \code{\link{CTSS}} object, for example extracted from a
\code{RangedSummarizedExperiment} object with the \code{\link{rowRanges}}
command.}

\item{annot}{A \code{\link{GRanges}} from which promoter positions will be inferred.
Typically GENCODE.  If the \code{type} metadata is present, it should
contain \code{gene}, \code{exon} and \code{transcript} among its values.  Otherwise,
all entries are considered transcripts. If the \code{transcript_type}
metadata is available, the entries that may not be primary products
(for instance \sQuote{snoRNA}) are discarded.}
}
\value{
A Run-length-encoded (\code{\link{Rle}}) factor of same length as the \code{CTSS}
object, indicating if the interval is \code{promoter}, \code{exon}, \code{intron} or
\code{unknown}, or just \code{promoter}, \code{gene}, \code{unknown} if the \code{type}
metadata is absent.
}
\description{
Assigns region types such as \code{promoter}, \code{exon} or \code{unknown}
to CTSSes.
}
\details{
Only the biotypes that are likely to have a promoter will be filtered
in.  This is currently hardcoded in the function; see its source code.  Example
of biotypes without a promoter: VDJ segments, etc.
}
\examples{
CAGEr:::ranges2annot(CTSScoordinatesGR(exampleCAGEexp), exampleZv9_annot)

}
\seealso{
\code{\link{CTSScoordinatesGR}}, \code{\link{exampleZv9_annot}}

Other CAGEr annotation functions: \code{\link{annotateCTSS}},
  \code{\link{plotAnnot}}, \code{\link{ranges2genes}},
  \code{\link{ranges2names}}
}
\author{
Charles Plessy
}
