% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotations.R
\name{ranges2annot}
\alias{ranges2annot}
\title{hierarchical annotation of CTSs}
\usage{
ranges2annot(ranges, annot)
}
\arguments{
\item{ranges}{The CTSS (GRanges) object, for example extracted from a
RangedSummarizedExperiment object with the \code{rowRanges}
command.}

\item{annot}{A \code{\link{GRanges}} from which promoter positions will
be inferred.  Typically GENCODE.  If the \code{type} metadata is present,
it should contain \sQuote{gene}, \sQuote{exon} and \sQuote{transcript} among
its values.  Otherwise, all entries are considered transcripts.
If the \code{transcript_type} metadata is available, the entries may
be not primary products (for instance \sQuote{snoRNA}) are discarded.}
}
\value{
A Run-length-encoded (Rle) factor of same length as the GRanges object,
        indicating if the interval is promoter, exon, intron or unknown, or
        just promoter, gene, unknown if the \code{type} metadata is absent.
}
\details{
Only the biotypes that are likely to have a promoter will be filtered
in.  This is currently hardcoded in the function; see its source code.  Example
of biotypes without a promoter: VDJ segments, etc.
}
\examples{
CAGEr:::ranges2annot(CTSScoordinatesGR(exampleCAGEexp), exampleZv9_annot)

}
\seealso{
\code{\link{CTSScoordinatesGR}}, \code{\link{exampleZv9_annot}}

Other CAGEr annotation functions: \code{\link{annotateCTSS}},
  \code{\link{plotAnnot}}, \code{\link{ranges2genes}}
}
\author{
Charles Plessy
}
