% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotations.R
\name{ranges2annot}
\alias{ranges2annot}
\title{ranges2annot}
\usage{
ranges2annot(ranges, annot, showClasses = NULL)
}
\arguments{
\item{ranges}{The Genomics Ranges object, for example extracted from a
RangedSummarizedExperiment object with the \code{rowRanges}
command.}

\item{annot}{A \code{\link{GRanges}} object containing \code{type} and
\code{transcript_type} metadata providing enough information for
inferring promoters, exons and introns.  Typically GENCODE.}
}
\value{
A Run-length-encoded (Rle) factor of same length as the GRanges object,
        indicating if the interval is promoter, exon, intron or unknown.
}
\description{
hierarchical annotation of a Genomic Range as promoter, exon or intron.
}
\examples{
\dontrun{
# With a CAGEexp object called "ce" and a GRanges object called "gff"
# containing transcript/exon information like with GENCODE:
gff <- readRDS(system.file("extdata/Zv9_annot.rds", package = "CAGEr"))
CTSScoordinatesGR(ce)$annotation <- ranges2annot(CTSScoordinatesGR(ce), gff)
colData(ce)[levels(CTSScoordinatesGR(ce)$annotation)] <-
  DataFrame(t(sapply(CTSStagCountDF(ce), function(X) tapply(X, CTSScoordinatesGR(ce)$annotation, sum))))
}

}
\seealso{
Other CAGEr annotation functions: \code{\link{annotateCTSS}},
  \code{\link{plotAnnot}}, \code{\link{ranges2genes}}
}
\author{
Charles Plessy
}
