% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotations.R
\name{ranges2genes}
\alias{ranges2genes}
\title{ranges2genes}
\usage{
ranges2genes(ranges, genes)
}
\arguments{
\item{ranges}{Genomics Ranges object, for example extracted from a
RangedSummarizedExperiment object with the \code{rowRanges}
command.}

\item{genes}{A \code{\link{GRanges}} object containing \code{gene_name} metadata.}
}
\value{
A character vector of same length as the GRanges object, indicating
        one gene symbol or a comma-separated list of gene symbols for each
        range.
}
\description{
Assign gene symbol(s) to Genomic Ranges.
}
\examples{
\dontrun{
# With a CAGEexp object called "ce" and a GRanges object called "gff"
# containing transcript/exon information like with GENCODE:
CTSScoordinatesGR(ce)$genes <- ranges2genes(CTSScoordinatesGR(ce), gff)
}

}
\seealso{
\code{\link{CTSScoordinatesGR}}

Other CAGEr annotation functions: \code{\link{annotateCTSS}},
  \code{\link{plotAnnot}}, \code{\link{ranges2annot}}
}
\author{
Charles Plessy
}
