% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAGEr.R, R/SetMethods.R
\docType{methods}
\name{sampleLabels}
\alias{sampleLabels}
\alias{sampleLabels}
\alias{sampleLabels<-}
\alias{sampleLabels<-,CAGEset}
\alias{sampleLabels<-,CAGEexp}
\alias{sampleLabels<-,CAGEer-method}
\title{Extracting CAGE datasets labels from CAGEr objects}
\usage{
sampleLabels(object)

sampleLabels(object) <- value
}
\arguments{
\item{object}{A CAGEr object.}

\item{value}{A character vector with a unique and valid name for each sample.
the \code{names} attributes indicate the colors.}
}
\value{
Returns a named character vector of labels of all CAGE datasets
present in the CAGEr object.  The values are the lables and the names
are the colors.
}
\description{
Extracts the labels and colors of CAGE datasets
from \code{\link{CAGEset}} and \code{\link{CAGEexp}} objects.
}
\details{
Renaming samples is possible only in \code{CAGEexp} objects, before
data is loaded.
}
\note{
If no colors are supplied, then default colors will be assigned
usign the \code{rainbow} function.  Assigned colors are not guaranteed
to be stable.
}
\examples{
load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
sampleLabels(exampleCAGEset)

}
\seealso{
\code{\link{setColors}}

Other CAGEr accessor methods: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScoordinates}},
  \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CTSSnormalizedTpm}},
  \code{\link{CTSStagCountTable}},
  \code{\link{CTSStagCount}}, \code{\link{genomeName}},
  \code{\link{inputFilesType}}, \code{\link{inputFiles}},
  \code{\link{librarySizes}}, \code{\link{tagClusters}}

Other CAGEr setter methods: \code{\link{genomeName}},
  \code{\link{inputFilesType}}, \code{\link{inputFiles}},
  \code{\link{setColors}}
}
\author{
Vanja Haberle

Charles Plessy
}
