% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\docType{methods}
\name{sampleLabels}
\alias{sampleLabels}
\title{Extracting CAGE datasets labels from CAGEr objects}
\usage{
sampleLabels(object)
}
\arguments{
\item{object}{A CAGEset or CAGEexp object.}
}
\value{
Returns a character vector of labels of all CAGE datasets present
in the CAGEr object.
}
\description{
Extracts the labels of CAGE datasets (samples, experiments)
from \code{\link{CAGEset}} and \code{\link{CAGEexp}} objects.
}
\examples{
load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
sampleLabels(exampleCAGEset)

}
\seealso{
Other CAGEr accessor methods: \code{\link{genomeName}},
  \code{\link{inputFilesType}}, \code{\link{inputFiles}},
  \code{\link{librarySizes}}
}
\author{
Vanja Haberle
}
