% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteringFunctions.R
\name{tagClusterConvertors}
\alias{tagClusterConvertors}
\alias{TCgranges2dataframe}
\alias{TCdataframe2granges}
\title{Private functions to convert TC formats}
\usage{
TCgranges2dataframe(gr)

TCdataframe2granges(df)
}
\arguments{
\item{gr}{Consensus clusters in \code{GRanges} format.}

\item{df}{Consensus clusters in \code{data.frame} format.}
}
\description{
Interconvert tag clusters (TC) formats used in classes CAGEset
(\code{data.frame}) and CAGEexp (\code{GRanges}).
}
\details{
The original format used in \code{\link{CAGEset}} objects follows BED
("0-based") conventsion for the start and end coordinates.  On the other
hand, the GRanges objects used in \code{\link{CAGEexp}} objects follow
the "1-based" convention.  Therefore a value of 1 has to be added or
subtracted to the start positions when converting between both formats.
}
\examples{
df <- tagClusters(exampleCAGEset, 1)
head(df)
gr <- CAGEr:::TCdataframe2granges(df)
gr
head(CAGEr:::TCgranges2dataframe(gr))
# No exact round-trip because start and end were not integer in df.
identical(df, CAGEr:::TCgranges2dataframe(gr))
if (! all(df == CAGEr:::TCgranges2dataframe(gr)))
  stop("No round-trip between TCdataframe2granges and TCgranges2dataframe")

tagClustersGR(exampleCAGEexp)
head(CAGEr:::TCgranges2dataframe(CAGEr:::tagClustersGR(exampleCAGEexp, 1)))
}
\seealso{
Other df2granges converters: \code{\link{consensusClusterConvertors}}
}
\concept{df2granges converters}
