% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTSS.R
\docType{class}
\name{TagClusters}
\alias{TagClusters}
\alias{.TagClusters}
\title{Tag clusters class}
\details{
The \code{TagClusters} class represents tag clusters.
It is used internally by CAGEr for type safety.
}
