% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\name{tagClusters}
\alias{tagClusters}
\title{Extract tag clusters (TCs) for individual CAGE experiments}
\usage{
tagClusters(object, sample = NULL, returnInterquantileWidth = FALSE,
  qLow = NULL, qUp = NULL)
}
\arguments{
\item{object}{A \code{\link{CAGEr}} object.}

\item{sample}{Label of the CAGE dataset (experiment, sample) for which to extract tag clusters.
If \code{sample = NULL}, a list of all the clusters for each samples is returned, and other
parameters are ignored.}

\item{returnInterquantileWidth}{Should the interquantile width for each tag cluster be returned.}

\item{qLow, qUp}{Position of which quantile should be used as a left (lower) or right (upper)
boundary (for \code{qLow} and \code{qUp} respectively) when calculating interquantile width. 
Default value \code{NULL} results in using the start coordinate of the cluster.  Used only
when \code{returnInterquantileWidth = TRUE}, otherwise ignored.}
}
\value{
Returns a \code{data.frame} with genomic coordinates, position of dominant TSS, total
CAGE signal and additional information for all TCs from specified CAGE dataset (sample).  If
\code{returnInterquantileWidth = TRUE}, interquantile width for each TC is also calculated
using specified quantile positions and returned in the data frame.
}
\description{
Extracts tag clusters (TCs) produced by \code{\link{clusterCTSS}} function for
a specified CAGE experiment from a CAGEr object.
}
\examples{
load(system.file("data", "exampleCAGEset.RData", package="CAGEr"))
TC <- tagClusters(object = exampleCAGEset, sample = "sample2", returnInterquantileWidth = TRUE, qLow = 0.1, qUp = 0.9)
head(TC)

ce <- readRDS(system.file(package = "CAGEr", "extdata/CAGEexp.rds"))
normalizeTagCount(ce)
clusterCTSS( object = ce, threshold = 50, thresholdIsTpm = TRUE
           , nrPassThreshold = 1, method = "distclu", maxDist = 20
           , removeSingletons = TRUE, keepSingletonsAbove = 100)
cumulativeCTSSdistribution(ce, clusters = "tagClusters")
quantilePositions( ce, clusters = "tagClusters"
                 , qLow = c(0.1,0.2), qUp = c(0.8,0.9))
TC <- tagClusters(object = ce, sample = "sample2", returnInterquantileWidth = TRUE, qLow = 0.1, qUp = 0.9)
head(TC)
}
\seealso{
Other CAGEr accessor methods: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScoordinates}},
  \code{\link{CTSSnormalizedTpm}},
  \code{\link{CTSStagCountTable}},
  \code{\link{CTSStagCount}}, \code{\link{genomeName}},
  \code{\link{inputFilesType}}, \code{\link{inputFiles}},
  \code{\link{librarySizes}}, \code{\link{sampleLabels}}

Other CAGEr clusters functions: \code{\link{CTSScumulativesTagClusters}},
  \code{\link{aggregateTagClusters}},
  \code{\link{clusterCTSS}},
  \code{\link{consensusClusters}},
  \code{\link{cumulativeCTSSdistribution}},
  \code{\link{plotInterquantileWidth}},
  \code{\link{quantilePositions}}
}
\author{
Vanja Haberle
}
