% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R, R/SetMethods.R
\name{tagClustersGR}
\alias{tagClustersGR}
\alias{tagClustersGR,CAGEexp-method}
\alias{tagClustersGR<-}
\alias{tagClustersGR<-,CAGEexp,ANY,TagClusters-method}
\alias{tagClustersGR<-,CAGEexp,missing,GRangesList-method}
\title{Extract tag clusters (TCs) for individual CAGE experiments}
\usage{
tagClustersGR(
  object,
  sample = NULL,
  returnInterquantileWidth = FALSE,
  qLow = NULL,
  qUp = NULL
)

\S4method{tagClustersGR}{CAGEexp}(
  object,
  sample = NULL,
  returnInterquantileWidth = FALSE,
  qLow = NULL,
  qUp = NULL
)

tagClustersGR(object, sample = NULL) <- value

\S4method{tagClustersGR}{CAGEexp,ANY,TagClusters}(object, sample = NULL) <- value

\S4method{tagClustersGR}{CAGEexp,missing,GRangesList}(object, sample = NULL) <- value
}
\arguments{
\item{object}{A \code{CAGEexp} object.}

\item{sample}{Label of the CAGE dataset (experiment, sample) for which to
extract tag clusters. If \code{samples = NULL}, a list of all the clusters for
each sample is returned.}

\item{returnInterquantileWidth}{Return the interquantile width for each tag cluster.}

\item{qLow, qUp}{Position of which quantile should be used as a left (lower)
or right (upper) boundary (for \code{qLow} and \code{qUp} respectively) when
calculating interquantile width.  Default value \code{NULL} results in using the
start coordinate of the cluster.  Used only when
\code{returnInterquantileWidth = TRUE}, otherwise ignored.}

\item{value}{A \code{\link{TagClusters}} object.}
}
\value{
Returns a \code{GRangesList} or a \code{GRanges} object with genomic coordinates,
position of dominant TSS, total CAGE signal and additional information for
all TCs from specified CAGE dataset (sample).  If
\code{returnInterquantileWidth = TRUE}, interquantile width for each TC is also
calculated using provided quantile positions.
}
\description{
Extracts tag clusters (TCs) produced by \code{\link{clusterCTSS}} function
for a specified CAGE experiment from a \code{\link{CAGEexp}} object.
}
\examples{
tagClustersGR( exampleCAGEexp, "Zf.high", TRUE, 0.1, 0.9 )
tagClustersGR( exampleCAGEexp, 1
             , returnInterquantileWidth = TRUE, qLow = 0.1, qUp = 0.9 )

}
\seealso{
Other CAGEr accessor methods: 
\code{\link{CTSSclusteringMethod}()},
\code{\link{CTSScoordinatesGR}()},
\code{\link{CTSScumulativesTagClusters}()},
\code{\link{CTSSnormalizedTpmDF}()},
\code{\link{CTSStagCountDF}()},
\code{\link{GeneExpDESeq2}()},
\code{\link{GeneExpSE}()},
\code{\link{consensusClustersGR}()},
\code{\link{expressionClasses}()},
\code{\link{genomeName}()},
\code{\link{inputFilesType}()},
\code{\link{inputFiles}()},
\code{\link{librarySizes}()},
\code{\link{sampleLabels}()},
\code{\link{seqNameTotalsSE}()}

Other CAGEr clusters functions: 
\code{\link{CTSSclusteringMethod}()},
\code{\link{CTSScumulativesTagClusters}()},
\code{\link{CustomConsensusClusters}()},
\code{\link{aggregateTagClusters}()},
\code{\link{clusterCTSS}()},
\code{\link{consensusClustersDESeq2}()},
\code{\link{consensusClustersGR}()},
\code{\link{cumulativeCTSSdistribution}()},
\code{\link{plotInterquantileWidth}()},
\code{\link{quantilePositions}()}
}
\author{
Vanja Haberle

Charles Plessy
}
\concept{CAGEr accessor methods}
\concept{CAGEr clusters functions}
