% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R, R/SetMethods.R
\docType{methods}
\name{tagClusters}
\alias{tagClusters}
\alias{tagClusters,CAGEr-method}
\alias{tagClustersGR}
\alias{tagClustersGR,CAGEset-method}
\alias{tagClustersGR,CAGEexp-method}
\alias{tagClustersGR<-}
\alias{tagClustersGR<-,CAGEset,ANY,ANY-method}
\alias{tagClustersGR<-,CAGEexp,ANY,TagClusters-method}
\alias{tagClustersGR<-,CAGEexp,missing,GRangesList-method}
\title{Extract tag clusters (TCs) for individual CAGE experiments}
\usage{
tagClusters(object, samples = NULL, returnInterquantileWidth = FALSE,
  qLow = NULL, qUp = NULL)

\S4method{tagClusters}{CAGEr}(object, samples = NULL,
  returnInterquantileWidth = FALSE, qLow = NULL, qUp = NULL)

tagClustersGR(object, sample = NULL, returnInterquantileWidth = FALSE,
  qLow = NULL, qUp = NULL)

\S4method{tagClustersGR}{CAGEset}(object, sample = NULL,
  returnInterquantileWidth = FALSE, qLow = NULL, qUp = NULL)

\S4method{tagClustersGR}{CAGEexp}(object, sample = NULL,
  returnInterquantileWidth = FALSE, qLow = NULL, qUp = NULL)

tagClustersGR(object, samples = NULL) <- value

\S4method{tagClustersGR}{CAGEset,ANY,ANY}(object,
  samples = NULL) <- value

\S4method{tagClustersGR}{CAGEexp,ANY,TagClusters}(object,
  samples = NULL) <- value

\S4method{tagClustersGR}{CAGEexp,missing,GRangesList}(object,
  samples = NULL) <- value
}
\arguments{
\item{object}{A \code{\link{CAGEr}} object.}

\item{samples}{Label of the CAGE dataset (experiment, sample) for which to extract tag clusters.
If \code{samples = NULL}, a list of all the clusters for each sample is returned.}

\item{returnInterquantileWidth}{Should the interquantile width for each tag cluster be returned.}

\item{qLow, qUp}{Position of which quantile should be used as a left (lower) or right (upper)
boundary (for \code{qLow} and \code{qUp} respectively) when calculating interquantile width.
Default value \code{NULL} results in using the start coordinate of the cluster.  Used only
when \code{returnInterquantileWidth = TRUE}, otherwise ignored.}

\item{sample}{Label of one CAGE dataset (experiment, sample) for which to extract tag
clusters. (For \code{tagClustersGR}, only one sample can be extracted.)}

\item{value}{A \code{\link{TagClusters}} object.}
}
\value{
Returns a \code{data.frame} with genomic coordinates, position of dominant TSS, total
CAGE signal and additional information for all TCs from specified CAGE dataset (sample).  If
\code{returnInterquantileWidth = TRUE}, interquantile width for each TC is also calculated
using specified quantile positions and returned in the data frame.
}
\description{
Extracts tag clusters (TCs) produced by \code{\link{clusterCTSS}} function for
a specified CAGE experiment from a CAGEr object.
}
\examples{
head(tagClusters( exampleCAGEset, "sample2"
                , returnInterquantileWidth = TRUE, qLow = 0.1, qUp = 0.9))
tagClustersGR(exampleCAGEexp, "Zf.high", TRUE, 0.1, 0.9)

}
\seealso{
Other CAGEr accessor methods: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScoordinates}},
  \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CTSSnormalizedTpm}},
  \code{\link{CTSStagCountTable}},
  \code{\link{CTSStagCount}}, \code{\link{GeneExpDESeq2}},
  \code{\link{GeneExpSE}},
  \code{\link{consensusClustersGR}},
  \code{\link{genomeName}}, \code{\link{inputFilesType}},
  \code{\link{inputFiles}}, \code{\link{librarySizes}},
  \code{\link{sampleLabels}}, \code{\link{seqNameTotalsSE}}

Other CAGEr clusters functions: \code{\link{CTSSclusteringMethod}},
  \code{\link{CTSScumulativesTagClusters}},
  \code{\link{CustomConsensusClusters}},
  \code{\link{aggregateTagClusters}},
  \code{\link{clusterCTSS}},
  \code{\link{consensusClustersDESeq2}},
  \code{\link{consensusClustersGR}},
  \code{\link{cumulativeCTSSdistribution}},
  \code{\link{plotInterquantileWidth}},
  \code{\link{quantilePositions}}
}
\author{
Vanja Haberle
}
\concept{CAGEr accessor methods}
\concept{CAGEr clusters functions}
