% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMethods.R
\name{tagClustersQuantile}
\alias{tagClustersQuantile}
\alias{tagClustersQuantile,TagClusters-method}
\alias{tagClustersQuantile,CAGEexp-method}
\title{Quantile metadata stored in CAGEr objects.}
\usage{
tagClustersQuantile(object, samples = NULL, q = NULL)

\S4method{tagClustersQuantile}{TagClusters}(object, samples = NULL, q = NULL)

\S4method{tagClustersQuantile}{CAGEexp}(object, samples = NULL, q = NULL)
}
\arguments{
\item{object}{A \code{\link{CAGEr}} object.}

\item{samples}{Sample name(s), number(s) or \code{NULL} (default) for all samples.}

\item{q}{A single quantile (not a list)}

\item{value}{A list (one entry per sample) of data frames with multiple columns:
\code{cluster} for the cluster ID, and then \code{q_0.n} where \code{0.n}
indicates a quantile.}
}
\value{
Returns a \code{data.frame} where the first column gives cluster
names and the next columns give quantile positions, in \emph{zero-based}
chromosome coordinates (because the tag clusters in former CAGEset objects were
represented in zero-based coordinates as well)).
}
\description{
Accessor functions to quantile metadata.
}
